==> ds/notebooks/01_data_exploration.ipynb <==
import re
import pandas as pd
import numpy as np
import PyPDF2

pd.set_option("display.max_columns", 85)
pd.set_option("display.max_rows", 85)

### NaturDoc - TL BL WT 22-23

# Data acquisition and exploration:

Initial datasets found by the pitcher:
* pay wall
* dataset of images for training scanning feature, but not a dataset containing information on symptoms, possible treatments etc. of herbs

We quickly realised that we would need to look for other datasets or find possible sources of scraping. We defined the one-sentence-pitch of our app as follows:

_Create an app to diagnose simple sickness based on form inputs, suggest nature-plant-based remedies or recipes to treat simple sickness._

Research into ethnobotany, folk medicine, notebooks on Kaggle etc. Some of the datasets/data sources we found were sometimes too simplistic, sometimes too specialised, sometimes in difficult to work formats. One dataset we found promising was the Duke Ethnobotanical dataset.

### Dr. Duke's Ethnobotanical Database:

[Dr. Duke's Phytochemical and Ethnobotanical Databases](https://catalog.data.gov/dataset/dr-dukes-phytochemical-and-ethnobotanical-databases-cecc0) contain a variety of databases, of which the ETHNOBOT.csv seemed particularly promising:

duke_df = pd.read_csv("../data/DUKE/ETHNOBOT.csv")

duke_df.head()

The dataset describes unique treatments by pairing an "activity" (= ailment targeted by this specific treatment) with a specific plant, further indicating the reference of the research that has described this relationship between plant and treatment and other information such as the country, genus and species, common name, family etc.

#### "Activities":

The "activities", i.e. curative properties ascribed to a given herb, cover a wide range of ailments and conditions.

activities = duke_df["ACTIVITY"].unique()
len(activities)

As we can see, over 2200 unique conditions are described by the dataset. Looking at these afflictions in more detail:

print(*activities[::25].tolist(), sep=" --- ")

The described disease/condition can range from a rather general description, such as aches, to specific ailments and injuries such as gunshot wounds, diabetes and so on. The category "spice" is also included. This wide range of descriptors warrants some further treatment of the data further down the line.

#### Herbs:

Herbs are mentioned by their taxonomic name and, if available, a common name. 
* Are there maybe several common names for the same herb mentioned at different data entries?

herbs = duke_df["TAXON"].unique()

len(herbs)

In total, the dataset contains data about over 13000 unique herbs that have been subjected to ethonobotanical or phytochemical research.

herbs_list = list()
for herb in duke_df.values:
    herbs_dict = dict()
    herbs_dict["TAXON"] = herb[14]
    herbs_dict["CNAME"] = herb[9]
    if herbs_dict not in herbs_list:
        herbs_list.append(herbs_dict)

Pairing of taxonomic name and common name (if available):

herbs_list[::250]
print(*herbs_list[::1000], sep=" --- ")

len(herbs_list)

For some reason, the length is longer than the count of unique herbs?
* because there can be several common names that will only be mentioned at different entries, so it creates a unique "TAXON" + "CNAME" dictionary in the list?

for herb in herbs_list:
    if herb["TAXON"] == "Crithmum maritimum":
        print(herb)

filt = (duke_df["TAXON"] == "Crithmum maritimum")
duke_df[filt]["CNAME"]

This need to be kept in mind further down the line when transforming data from the dataset.
* create dataframe sorted by unique herb (== "TAXON"), containing lists of all "CNAMES" and "ACTIVITIES" 
    * ignore rest of the data?

##### Ex. Ginger:

filt_ginger = (duke_df["CNAME"] == "Ginger")
duke_df[filt_ginger]

## Royal Botanic Gardens Kew Datasets:

[Plants of the World Online by the Royal Botanic Gardens, Kew](https://powo.science.kew.org/)

_Note: some of the csv files from the KEW datasets were too large to be pushed to Github, so please be aware that it's best to leave the cells with their output as they are and not execute them again. In the next notebook, we will include the two csv files that contain the most useful information for us, reduced in size._

A really nice and thorough website, however downloadable csv's do not contain all the information from the website. Some of the datasets contain the links to the webpages for each individual plant, however a more thorough look reveals that there's no standardised presentation for the plants. Therefore, scraping for information from the website seems futile.

#### DWCA:

dwca_dis_df = pd.read_csv("../data/DWCA/wcvp_distribution.csv", sep="|")
dwca_rep_df = pd.read_csv("../data/DWCA/wcvp_replacementNames.csv", sep="|")

The wcvp_taxon.csv is extremely large (not pushed to github).

dwca_tax_df = pd.read_csv("../data/DWCA/wcvp_taxon.csv", sep="|")

### Head:

Something that these datasets might be useful for when cross-referencing is their information of localisation of the plants. Warrants more thorough examination.

dwca_dis_df.head()

dwca_rep_df.head()

dwca_tax_df.head()

"taxonremarks" to aid in localising the plants?

len(dwca_tax_df["taxonremarks"].unique().tolist())

dwca_val = dwca_tax_df["taxonremarks"].value_counts()

print(dwca_val[::4000])

Too many values of inconsistent format?

len(dwca_dis_df["locality"].unique())

dwca_dis_df["locality"].value_counts()[::10]


Less values but also somewhat inconsistent format? 

#### WCVP:

Both of these csv files are too large to be pushed to GitHub (e.g. wcvp_distribution.csv has over 1 million entries):

wcvp_dis_df = pd.read_csv("../data/WCVP/wcvp_distribution.csv", sep="|")
wcvp_names_df = pd.read_csv("../data/WCVP/wcvp_names.csv", sep="|")

Contains continent, region and area - seems promising?

wcvp_dis_df.head()

wcvp_names_df.head()

### Crossreferencing DUKE and KEW datasets:

duke_df.loc[:, "TAXON"]

dwca_tax_df.loc[:, "scientfiicname"]

DUKE dataset filtered by matches scientific names matching those found in the KEW dataset:

filter = duke_df.loc[:, "TAXON"].isin(dwca_tax_df.loc[:, "scientfiicname"])
filt_df = duke_df[filter]
filt_df.shape

len(duke_df) - len(filt_df)

8232 entries were thus discarded from the DUKE dataset.

(Keep in mind, every singular entry is a combination of a plant with a specific activity, so this number does not tell us how many plants were dropped.)

print(len(duke_df["TAXON"].unique()))
print(len(filt_df["TAXON"].unique()))
len(duke_df["TAXON"].unique()) - len(filt_df["TAXON"].unique())


Entries belonging to 2384 unique scientific plant names were dropped from the DUKE dataset.

## WHO Monographs:

[WHO monographs on selected medicinal plants](https://apps.who.int/iris/handle/10665/42052)

There are several WHO monographs describing herbs and their medicinal applications in depth. The format presents some challenges, however the description of each herb follows the same structure. The text itself is written free-form and some information is therefore difficult to extract. Regardless, these monographs seem promising for the purpose of enriching the DUKE dataset with some more detailed text, e.g. serving as "instructions" of sorts.

Accessing a pdf with PyPDF2:

# creating a pdf file object
who_01_pdf = open('../data/WHO/monograph_01.pdf', 'rb')
  
# creating a pdf reader object
who_01_reader = PyPDF2.PdfReader(who_01_pdf)
  
# printing number of pages in pdf file
print(len(who_01_reader.pages))
  
# creating a page object
page_obj = who_01_reader.pages[4]
  
# extracting text from page
print(page_obj.extract_text())
  
# closing the pdf file object
who_01_pdf.close()

!!! For this specific pdf, page order is mangled:
Page 1 is the 8th actual page (+ 7)
* OK up to page 50 
* 59, 60 
* 67-76 
* 61-66
* 51-58
* 77-297

Issues with weird characters, e.g. when writing "fi" (rendered identically in markdown) or "-":

-> <code>"Rauwol ﬁae"</code> Rauwolfiae

-> <code>"Deﬁnition"</code> Definition

sometimes adds a space, but inconsistent?

#### Reading and cleaning up the pdf:

Regex used to target unusual "fi" and "-" characters, citations such as "( 1, 2)", spaces before dots. Every page is appended to a list of pages, so that I no longer need to open/read the pdf all over again.

who_01_pdf = open('../data/WHO/monograph_01.pdf', 'rb')

who_01_reader = PyPDF2.PdfReader(who_01_pdf)

who_01_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_spacequotesingle = re.compile(" ’")
pattern_spacequotedouble = re.compile(" ”")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_01_reader.pages)):
    page_obj = who_01_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_spacequotesingle, "'", text)
    text = re.sub(pattern_spacequotedouble, "\"", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_01_pages.append(lines)
    
who_01_pdf.close()

Creating slices of wrong page order:

p_01_50 = who_01_pages[7 : 57]
p_51_58 = who_01_pages[75 : 83]
p_59_60 = who_01_pages[57 : 59]
p_61_66 = who_01_pages[69 : 75]
p_67_76 = who_01_pages[59 : 69]
p_77_end = who_01_pages[83 : ]

Appending slices to a new list:

who_01_pages_corrected = p_01_50 + p_51_58 + p_59_60 + p_61_66 + p_67_76 + p_77_end

who_01_pages_corrected[50][:10]

As we can see, page 51 is now at the correct index position (== 50, as indices start from 0).

Compared to previous list, page 51 is at index 75:

(Also need to always add 7 to index as document includes cover, starts counting from i to vi before starting over at page 1.)

who_01_pages[75][:10]

#### WHO Remedies:
These already describe the specific name, i.e. which part of plant used:

for page in who_01_pages_corrected:
    for i, line in enumerate(page):
        if "Definition" == line:
            output = re.sub(r"[0-9]*", "", page[i-1])
            print(output)

### Definition:

For example, _Bulbus Allii Cepae_ specifically describes the fresh or dried bulbs of onions (taxonomic name _Allium cepa_). The name of the plant itself is mentioned following the definition header in the monograph:

who_01_def = list()
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Definition" == line:
            n_lines = 1
            while page[i + n_lines] != "Synonyms" and i + n_lines < max_len:
                n_lines += 1
            who_01_def.append(" ".join(page[i + 1 : i + n_lines]))

print(*who_01_def[:5], sep="\n")

It is therefore possible to cross-reference the WHO definition with the scientific names saved in the DUKE dataset (see later chapter).

### Geographical distribution:

Issue: Format of information varies a lot. 

Described with continents, countries and regions. Difficult to cross-reference. 

Different information about: native country, spread, commercial availability...

who_01_dis = list()
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Geographical distribution" == line:
            n_lines = 1
            while page[i + n_lines] != "General identity tests" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])

            who_01_dis.append(output)

print(*who_01_dis[:5], sep="\n")

### Vernacular names:

Generally follows the format of "name"-"comma"-"space"-"name"-"comma"-"space". Sometimes includes a line of text as first sentence which ends with a full-stop.

Try splitting at dot, if length of returned list is longer than 2 remove index 1?

Later split at "comma"-"space" to return a list of names?

who_01_vernacular = list()
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Selected vernacular names" == line:
            n_lines = 1
            while page[i + n_lines] != "Description" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])

            who_01_vernacular.append(output)

print(*who_01_vernacular[:5], sep="\n")

### Medicinal uses:

#### Clinical:

Issues with chemical illustrations being "read" at the end of the page, such as:

.chamazuleneCH3 CH3CH3 (/H11002)αbisabololCH3OH CH3H CH3CH3 O OOH OHOH apigenin --

who_01_clinical = list()
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" == line:
            n_lines = 1
            while page[i + n_lines] != "Uses described in pharmacopoeias and in traditional systems of" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])

            who_01_clinical.append(output)

who_01_clinical[8]

Regex to target chemical formulae: they appear right after the final dot without a space in between:

who_01_clinical = list()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" == line:
            n_lines = 1
            while page[i + n_lines] != "Uses described in pharmacopoeias and in traditional systems of" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            who_01_clinical.append(output)

who_01_clinical[8]

print(*who_01_clinical[:5], sep="\n")

#### Traditional:

Include same regex pattern:

who_01_traditional = list()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in pharmacopoeias and in traditional systems of" == line:
            n_lines = 2
            while "Uses described in folk medicine" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            who_01_traditional.append(output)

print(*who_01_traditional[:5], sep="\n")

#### Folk medicine:

who_01_folk = list()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")
  
for page in who_01_pages_corrected:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in folk medicine" in line:
            n_lines = 2
            while page[i + n_lines] != "Pharmacology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            who_01_folk.append(output)

print(*who_01_folk[:5], sep="\n")

Other areas of interest include:
* __Posology__
    * Posology, Dosage Forms
* __Toxicology__
* __Contraindications__
    * Contraindications, Warnings, Precautions, Adverse reactions

The _Precautions_ as described by the WHO monographs can have many different types of parameters, i.e. _General_, _Drug interactions_, _Carcinogenesis_ etc. Accordingly, it is difficult to account for all possible categories, not to mention that this is probably more information than we would like to present in our website, regardless.

### Issues with wrong or missing matches as well as "null" values:

As we can see here, the lengths of the various lists do not match and the information is therefore all mixed up:

print(len(who_01_def))
print(len(who_01_dis))
print(len(who_01_vernacular))
print(len(who_01_clinical))
print(len(who_01_traditional))
print(len(who_01_folk))

Upon further testing the data extracted so far, we realised that footnotes cause issues (similar to chemical formulae). When the pdf is read, they are placed at the beginning of each page. This results in some matches being missed when comparing with an equality operator; the algorithm checking for matching headings should instead use the "in" keyword.

#### Rerunning code with "in"-keyword:

When rerunning the previous extraction code with the "in" keyword, we sometimes need to make sure that words like "see" or "and" are not included to avoid wrong matches.

Skipping "distribution", as it is difficult to use and the KEW datasets as well as our pytrend endeavors provide better data.

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")

# Definitions with in-keyword:

who_01_def = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Definition" in line:
            n_lines = 1
            while page[i + n_lines] != "Synonyms" and i + n_lines < max_len:
                n_lines += 1
            who_01_def.append(" ".join(page[i + 1 : i + n_lines]))

# Vernacular names with in-keyword:

who_01_vernacular = list()
reg_pattern = re.compile(r"\.")
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Selected vernacular names" in line:
            n_lines = 1
            while page[i + n_lines] != "Description" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])
            
            output =  output.replace(".", "")
            
            # Sometimes taxonomic names appear in the WHO document names as a header of sorts:
            for name in duke_names:
                output = output.replace(name, ",")
            
            output = output.split(", ")

            who_01_vernacular.append(output)

# Clinical medicine with in-keyword:

who_01_clinical = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" in line:
            n_lines = 1
            while "Uses described in pharmacopoeias and in traditional" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            # in case the description gets cut off, drop the incomplete sentence:
            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."
            
            who_01_clinical.append(output)

# Traditional medicine with in-keyword:

who_01_traditional = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in pharmacopoeias and in traditional" in line:
            n_lines = 2
            while "Uses described in folk medicine" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_traditional.append(output)

# Folk with in-keyword:

who_01_folk = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in folk medicine" in line:
            n_lines = 2
            while page[i + n_lines] != "Pharmacology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_folk.append(output)

Verifying identical lengths:

print(len(who_01_def))
print(len(who_01_vernacular))
print(len(who_01_clinical))
print(len(who_01_traditional))
print(len(who_01_folk))

### Completing the lists for remaining values:

Add posology, adverse reactions, contraindication and warnings:

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")
            
who_01_con = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Contraindications" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Warnings" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_con.append(output)
            
who_01_warn = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Warnings" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Precautions" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_warn.append(output)
            
who_01_adv = list()

ref_pattern = re.compile(r"[0-9]+\.")  

for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Adverse reactions" in line and "see" not in line.lower() and not re.search(ref_pattern, line):
            n_lines = 1
            while page[i + n_lines] != "Posology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_adv.append(output)

who_01_pos = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Posology" in line and "and" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "References" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                # output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                # output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_pos.append(output)

Verifying the lengths of these new lists:

print(len(who_01_con))
print(len(who_01_warn))
print(len(who_01_adv))
print(len(who_01_pos))

### Null values:

The WHO monograph sometimes contains strings that indicate that there is no information available about something for a specific herb. These can be considered as "null" values of sorts.

who_01_warn[6:13]

As we can see, there are different sentences describing this absence of information. For now, we think it is best to replace them with an empty string.

null_values = ["No information available.", "None."]

def make_nan(target_list: list):
    for i, value in enumerate(target_list):
        if value in null_values:
            target_list[i] = ""

make_nan(who_01_clinical)
make_nan(who_01_traditional)
make_nan(who_01_folk)
make_nan(who_01_con)
make_nan(who_01_warn)
make_nan(who_01_adv)
make_nan(who_01_pos)

who_01_warn[11:13]

## Looking Ahead:

In the next notebook, we take a closer look at the datasets explored here to identify and experiment with possible approaches of manipulating and extracting data.
