==> ds/notebooks/02_a_data_transformation.ipynb <==
import re
import pandas as pd
import numpy as np
import PyPDF2

### NaturDoc - TL BL WT 22-23

# Data transformation:

Datasets we have explored:
* not useable by themselves
* not matching visions of team for information that the app should provide

So we delve deeper into the datasets and try to manipulate them to extract the information we need, in hopes of creating a final enriched dataset at the end.

## Imports:

### Duke:

duke_df = pd.read_csv("../data/DUKE/ETHNOBOT.csv")

### KEW:

The two KEW datasets with the most promising geographic data.

_Note: both of these csv files were too large to be pushed to GitHub. We will modify them in this notebook to export to csv with a reduced file size and make them more useable for us as well, as some code can be very time-consuming to run with the way the files are set up right now._

wcvp_dis_df = pd.read_csv("../data/WCVP/.wcvp_distribution.csv", sep="|")
wcvp_names_df = pd.read_csv("../data/WCVP/.wcvp_names.csv", sep="|")

### WHO Monograph 1:

who_01_pdf = open('../data/WHO/monograph_01.pdf', 'rb')

who_01_reader = PyPDF2.PdfReader(who_01_pdf)

who_01_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_spacequotesingle = re.compile(" ’")
pattern_spacequotedouble = re.compile(" ”")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_01_reader.pages)):
    page_obj = who_01_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_spacequotesingle, "’", text)
    text = re.sub(pattern_spacequotedouble, "”", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_01_pages.append(lines)
    
who_01_pdf.close()

p_01_50 = who_01_pages[7 : 57]
p_51_58 = who_01_pages[75 : 83]
p_59_60 = who_01_pages[57 : 59]
p_61_66 = who_01_pages[69 : 75]
p_67_76 = who_01_pages[59 : 69]
p_77_end = who_01_pages[83 : ]

who_01_pages_corrected = p_01_50 + p_51_58 + p_59_60 + p_61_66 + p_67_76 + p_77_end

## Transforming the DUKE dataset:

The original Dataset of the _Dr. Duke's Phytochemical and Ethnobotanical Databases_ contains a lot of information that we do not need for the scope of our app. Currently the dataset pairs unique herbs with a single "activity", linked to relevant research and other data, but for our needs it makes sense to pivot the dataframe by each herb, aggregating their data all in a single row.

As np.NaN values cause issues when joining the aggregated values subsequently, all np.NaN values will first be replaced with a custom nan string value.

target_cols = ["TAXON", "ACTIVITY", "CNAME", "FAMILY", "GENUS", "SPECIES"]
mod_duke_df = duke_df[target_cols].replace(np.NaN, "nan").pivot_table(index='TAXON', aggfunc=pd.unique)
mod_duke_df.head()

While multiple values are stored in a list for now, exporting and later re-importing this dataframe as a csv causes issues (the list will instead be treated as a string). This will be further explored in the "Pivoting" as well as the "Comparing output of these new csv files" subchapters of the section transforming the KEW datasets.

The following code instead joins the multiples values in a comma separated string:

mod_duke_df["ACTIVITY"] = mod_duke_df["ACTIVITY"].str.join(",")
mod_duke_df["CNAME"] = mod_duke_df["CNAME"].str.join(",")
mod_duke_df["FAMILY"] = mod_duke_df["FAMILY"].str.join(",")
mod_duke_df["GENUS"] = mod_duke_df["GENUS"].str.join(",")
mod_duke_df["SPECIES"] = mod_duke_df["SPECIES"].str.join(",")
mod_duke_df.head()

_Note: later, it is important to remember to remove the null values when they appear as part of a list._

The dataframe is now indexed by the unique taxonomic name. If necessary, a regular serial index can be created by executing reset_index(). This is helpful in that "TAXON" will then become a normal column again and can therefore be targeted with .loc.

_Edit: when exporting the dataframe to a new CSV, another serial index gets added upon reading it again later so this is not necessary._

mod_duke_df = mod_duke_df.reset_index()
mod_duke_df.head()

Exporting the transformed dataset for later use:

# mod_duke_df.to_csv("../data/DUKE/PIVOT_ETHNOBOT.csv")

### Targeting only the activities to create a placeholder collection of symptoms:

symptoms = duke_df["ACTIVITY"].unique()
symptoms_df = pd.DataFrame(symptoms)
symptoms_df.rename(columns={0: "symptomName"}, inplace=True)
filt = (symptoms_df.symptomName.str.contains("(Veterinary)") == False)
reduced_symptoms_df = symptoms_df[filt]
reduced_symptoms_df.head()

For placeholder json: requires ids

id_df = reduced_symptoms_df.copy()
reduced_symptoms_df.loc[:, "_id"] = id_df.index
reduced_symptoms_df.head()

reduced_symptoms_df.to_json("../output/symptoms.json", orient="records")

check_df = pd.read_json("../output/symptoms.json")
check_df.head()

## Transforming the KEW dataset:

### wcvp_names.csv:

wcvp_names_df.shape

As we can see, the wcvp_names_df has 1.414.666 rows of data and 31 columns. 

wcvp_names_df.head()

Taking a look at the contents of the dataframe, the following transformations seem plausible:
* columns we care about are plant_name_id and the taxonomic name so that we can properly identify the plants from the KEW datasets
* the existing taxon_name column contains information on hybridisation and so on, so it makes sense to create a new, simplified taxon column from genus and species

#### New "taxon" column:

wcvp_names_df["taxon"] = wcvp_names_df["genus"] + " " + wcvp_names_df["species"].str.lower()

wcvp_names_df.loc[:9, "taxon"]

#### Reducing dataframe to only the columns we need:

reduced_names_df = wcvp_names_df[["plant_name_id", "taxon"]].set_index("plant_name_id")
reduced_names_df

#### Exporting a new, lightweight csv file to upload to GitHub and use for further manipulations:

reduced_names_df.to_csv(r"../data/WCVP/REDUCED_wcvp_names.csv")

### wcvp_distribution.csv

wcvp_dis_df.shape

As we can see, the wcvp_dis_df has an astounding 1.950.339 rows of data for 11 columns. 

wcvp_dis_df.head()

Taking a look at the contents of the dataframe, the following transformations seem plausible:
* columns we care about are plant_name_id (for matching to the names dataframe) and continent, region, area (although not all three are necessary)
* instead of unique pairings of plant_locality_id and plant_name_id, we can pivot the dataframe by unique plant_name_id instead

#### Reducing dataframe to only the columns we need:

reduced_dis_df = wcvp_dis_df[["plant_name_id", "continent", "region", "area"]]
reduced_dis_df

In testing, this has already brought the file size from 138MB down to 106MB.

#### Pivoting:

pivot_dis_df = reduced_dis_df.pivot_table(index='plant_name_id', aggfunc=pd.unique)
pivot_dis_df

1,950,339 rows down to 435,009 rows.

In testing, this brought the file size further down to 58MB (which is still too large for default GitHub uploads). Upon further consideration, the region column in particular seems the least useful and will therefore also be excluded in the final transformation:

reduced_dis_df = wcvp_dis_df[["plant_name_id", "continent", "area"]]
pivot_dis_df = reduced_dis_df.pivot_table(index='plant_name_id', aggfunc=pd.unique)
pivot_dis_df

# pivot_dis_df.to_csv(r"../data/WCVP/REDUCED_PIVOT_wcvp_distribution.csv")

##### Alternative to storing an array: 

Instead of aggregating values in a list (which causes issues with how the csv will be read later on), save it in a comma separated string. 

This throws an error and causes the area column to be dropped:

_['area'] did not aggregate successfully. If any error is raised this will raise in a future version of pandas. Drop these columns/ops to avoid this warning._

agg_func = lambda x: ",".join(x.unique())

# pivot_dis_df = reduced_dis_df.pivot_table(index='plant_name_id', aggfunc=agg_func)
# pivot_dis_df

##### Alternative 2 to storing an array: 

Transform values saved in pivot_dis_df after the fact:

pivot_dis_df_copy = pivot_dis_df.copy()
pivot_dis_df_copy["continent"] = pivot_dis_df_copy["continent"].str.join(",")
pivot_dis_df_copy["area"] = pivot_dis_df_copy["area"].str.join(",")
pivot_dis_df_copy

# pivot_dis_df_copy.to_csv(r"../data/WCVP/STR_REDUCED_PIVOT_wcvp_distribution.csv")

This actually further reduced the file size to 31MB.

#### Comparing output of these new csv files:

dis_df_a = pd.read_csv("../data/WCVP/REDUCED_PIVOT_wcvp_distribution.csv", sep=",")
dis_df_b = pd.read_csv("../data/WCVP/STR_REDUCED_PIVOT_wcvp_distribution.csv", sep=",")

Looking at the dataframes, the one generated from the initial approach seems to store the area in a [] list with each individual string surrounded by single quotation marks (spoiler: it's actually not a list/array).

The alternative approach of transforming entries from a list into a joined string stores the values separated by commas, without any brackets and quotation marks appearing at all.

dis_df_a.head()

dis_df_b.head()

When we take a look at an individual entry, however, we can see why the alternative approach is a lot more useable. The "list" of the initial approach is not a list at all, but rather a string containing brackets and quotation marks. __Extracting individual entries in a useable way therefore requires a lot of extra steps.__

print(type(dis_df_a.loc[0, "area"]))

# trying to store individual entries in an actual python list:
a_values = dis_df_a.loc[0, "area"]
a_values = a_values.split("'")

# splitting like this leaves nonsensical entries in the list, such as '[', ' ', 
# and even '\n '
bad_entries = ["[", "]", " ", "\n "]
for entry in bad_entries:
# recursive, since " " appears several times
    while entry in a_values:
        a_values.remove(entry)

print(a_values)
print(type(a_values))

Compared to extracting a useable list from the alternative csv file:

print(type(dis_df_b.loc[0, "area"]))

b_values = dis_df_b.loc[0, "area"]
b_values = b_values.split(",")

print(b_values)
print(type(b_values))

## Extracting useable data from the WHO monograph:

### Reading the Monograph and generating lists, as per the data exploration notebook:

duke_names = duke_df["TAXON"].unique().tolist()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")

# Definitions with in-keyword:

who_01_def = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Definition" in line:
            n_lines = 1
            while page[i + n_lines] != "Synonyms" and i + n_lines < max_len:
                n_lines += 1
            who_01_def.append(" ".join(page[i + 1 : i + n_lines]))

# Vernacular names with in-keyword:

who_01_vernacular = list()
reg_pattern = re.compile(r"\.")
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Selected vernacular names" in line:
            n_lines = 1
            while page[i + n_lines] != "Description" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])
            
            output =  output.replace(".", ",")
            
            # Sometimes taxonomic names appear in the WHO document names as a header of sorts:
            for name in duke_names:
                output = output.replace(name, ",")
            
            output =  output.replace(", ", ",")

            who_01_vernacular.append(output)

# Clinical medicine with in-keyword:

who_01_clinical = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" in line:
            n_lines = 1
            while "Uses described in pharmacopoeias and in traditional" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            # in case the description gets cut off, drop the incomplete sentence:
            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."
            
            who_01_clinical.append(output)

# Traditional medicine with in-keyword:

who_01_traditional = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in pharmacopoeias and in traditional" in line:
            n_lines = 2
            while "Uses described in folk medicine" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_traditional.append(output)

# Folk with in-keyword:

who_01_folk = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in folk medicine" in line:
            n_lines = 2
            while page[i + n_lines] != "Pharmacology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_folk.append(output)

who_01_con = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Contraindications" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Warnings" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_con.append(output)
            
who_01_warn = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Warnings" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Precautions" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_warn.append(output)
            
who_01_adv = list()

ref_pattern = re.compile(r"[0-9]+\.")  

for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Adverse reactions" in line and "see" not in line.lower() and not re.search(ref_pattern, line):
            n_lines = 1
            while page[i + n_lines] != "Posology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_adv.append(output)

who_01_pos = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Posology" in line and "and" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "References" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                # output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                # output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_pos.append(output)

Replace WHO null values with empty string:

null_values = ["No information available.", "None."]

def make_nan(target_list: list):
    for i, value in enumerate(target_list):
        if value in null_values:
            target_list[i] = ""

make_nan(who_01_clinical)
make_nan(who_01_traditional)
make_nan(who_01_folk)
make_nan(who_01_con)
make_nan(who_01_warn)
make_nan(who_01_adv)
make_nan(who_01_pos)

## Cross-reference DUKE:

It is possible to compare entries to DUKE dataset. When it is a match, we append a dictionary with name, definition etc. keys for the matched herb to a list.

The idea is to eventually add columns to the DUKE dataframe that includes some of this information.

### List of dictionaries:

who_01_remedies = list()

for i, definition in enumerate(who_01_def):

    remedy_dict = dict()

    for name in duke_names:
        if name in definition:
            remedy_dict["NAME"] = name
            remedy_dict["DEF"] = definition
            remedy_dict["VERNAC"] = who_01_vernacular[i]
            remedy_dict["CLINICAL"] = who_01_clinical[i]
            remedy_dict["TRADITIONAL"] = who_01_traditional[i]
            remedy_dict["FOLK"] = who_01_folk[i]

    if remedy_dict:
        who_01_remedies.append(remedy_dict)


#### Missing herbs:

We know that all the lists now have a length of 28 entries, however the dictionary created only contains 26:

print("Length of a list:", len(who_01_def), "\nLength of the dictionary:", len(who_01_remedies))

We will run a quick algorithm to see which two herbs were not included because their names did not return a match with the current dictionary generation code:

matches = list()

for definition in who_01_def:
    for herb in who_01_remedies:
        if definition in herb["DEF"]:
            matches.append(definition)

rest = list()

for definition in who_01_def:
    if definition not in matches:
        rest.append(definition)

print("Length of matches:", len(matches), "\nLength of rest:", len(rest), end="\n\n")
print(*rest, sep="\n")

Excluded herbs were:
* Chamomilla recutita
* Plantago afra

Filtering the Duke dataframe for "Chamomilla":

filt = (duke_df["TAXON"].str.contains("Chamomilla"))
duke_df[filt]["TAXON"]

Trying the lowercase "chamomilla" instead:

filt = (duke_df["TAXON"].str.contains("chamomilla"))
duke_df[filt]["TAXON"].unique()

From __Wikipedia, The Free Encyclopedia__: "Matricaria chamomilla (synonym: Matricaria recutita), commonly known as chamomile"

The WHO monograph and Duke's Phytochemical and Ethnobotanical Database refer to the same herb by different names.

Filtering the Duke dataframe for "Plantago afra":

filt = (duke_df["TAXON"].str.contains("Plantago afra"))
duke_df[filt]["TAXON"].unique()

Trying just "Plantago" instead to find similar matches:

filt = (duke_df["TAXON"].str.contains("Plantago"))
duke_df[filt]["TAXON"].unique()

From its definition, we know that "Plantago afra" has the following synonyms: 

"_Plantago afra L., P. indica L., P. ovata Forsk., or P. asiatica_"

As we can see in the above array returned by the Duke dataframe, both _Plantago ovata_ and _Plantago asiatica_ are in fact part of the Duke dataset.

#### Conclusion:

Both _Chamomilla recutita_ and _Plantago afra_ will have to have their entries corrected manually.

For _Plantago afra_, maybe a solution when initially checking for the name: check if the split name is in the definition.

matches = list()

for definition in who_01_def:
    for name in duke_names:
        name = name.split(" ")

        # some taxonomic names have an "x", indicating hybrids
        if "x" in name: 
            name.remove("x")

        if name[0] in definition and name[1] in definition:
            matches.append(" ".join(name))

print(*set(matches), sep=" - ")
len(matches)

What is especially interesting to see is that we now actually have 44 matches.

The definition often includes synonyms (such as we saw with _Plantago Afra_) so the subsequent information is relevant for a number of herbs in the Duke dataset. 

for match in matches:
    if "Plantago" in match:
        print(match)

But still no matches for "Chamomille", as its definition did not include a synonym that matched with the Duke dataset:

for match in matches:
    if "chamomilla" in match:
        print(match)

### Final dictionary generation code:

who_01_remedies = list()

for i, definition in enumerate(who_01_def):

    for name in duke_names:

        remedy_dict = dict()
        name = name.split(" ")

        # some taxonomic names have an "x", indicating hybrids
        if "x" in name: 
            name.remove("x")

        if name[0] in definition and name[1] in definition:
            remedy_dict["NAME"] = " ".join(name)
            remedy_dict["DEF"] = definition
            remedy_dict["VERNAC"] = who_01_vernacular[i]
            remedy_dict["CLINICAL"] = who_01_clinical[i]
            remedy_dict["TRADITIONAL"] = who_01_traditional[i] 
            remedy_dict["FOLK"] = who_01_folk[i]
            remedy_dict["CONTRAINDICATION"] = who_01_con[i]
            remedy_dict["WARNING"] = who_01_warn[i]
            remedy_dict["ADVERSE"] = who_01_adv[i]
            remedy_dict["POSOLOGY"] = who_01_pos[i]

        if remedy_dict:
            who_01_remedies.append(remedy_dict)

len(who_01_remedies)

print(who_01_remedies[0])

This code will be used in the next notebook as a basis to enrich the Duke dataframe.

Output name of herbs

mod_duke_df["popularityRating"] = np.nan

mod_duke_df.rename(columns={
    "TAXON": "taxonomicName",
    "ACTIVITY": "activity",
    "CNAME": "commonNames"
    }, inplace=True)
mod_duke_df

taxonomy_df = mod_duke_df[["taxonomicName", "popularityRating", "commonNames"]]
taxonomy_df.set_index("taxonomicName", inplace=True)
taxonomy_df.head()

taxonomy_df.commonNames = taxonomy_df.commonNames.str.replace('nan,', '', regex=True)

nan_toreplace = ["nan", ""]
taxonomy_df = taxonomy_df.replace(nan_toreplace, np.nan)

taxonomy_df.dropna(subset="commonNames")

filter = taxonomy_df.commonNames.str.contains('Ginger')
filt_df = taxonomy_df[filter]
filt_df

taxonomy_df.head()

taxonomy_df.to_csv("../output/taxonomic_names.csv", sep=",")

test_df = pd.read_csv("../output/taxonomic_names.csv")

test_df

## Looking Ahead:

The next few notebooks will explore a different avenue to supplement our data: using the pytrends module to pull popularity data for herbs from Google Trends, in the hopes of using that data as a general popularity indicator of a given herb. For now, we will not worry about getting region-specific data.
