==> ds/notebooks/02_b_Popular.Herbs.Duke.ipynb <==
 Our focus is to find the most popular herbs from the . In this case we used  pytrends.
 
 

import os
#os.chdir('pytrends/')

import sys
sys.path.insert(0,os.getcwd())

print(os.getcwd())
print(sys.path)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import random
from typing import List, Dict

from pytrends.request import TrendReq

from tqdm import tqdm
from itertools import islice
from IPython.display import display
import time

from warnings import warn

**Pytrend_score_group**
This function takes a list of terms, builds a payload with the TrendReq object from the pytrends library nd retrieves the average interest over time for each term. The function returns a dictionary of the ranks of the terms based on their interest over time.

def pytrend_score_group(terms : list, print_table : bool = False) -> Dict[str,int]:
    pt = TrendReq(retries=5)
    pt.build_payload(terms)
    tmp_df = pt.interest_over_time()
    tmp_df = tmp_df[terms].mean()
    
    if print_table:
        display(tmp_df)
        
    ranks = tmp_df.rank().to_dict()
    return ranks 

def break_into_groups(my_list : list, times : int):
    """times is the size of the group"""
    sublists = []
    for i in range(0, len(my_list), times):
        sublists.append(my_list[i:i+5])
    return sublists

def rank_by_popularity(group: List[List[str]], debug = True) -> List[Dict[str, int]]:
    ranked_groups = []
    for herbs in group:
        ranked_herbs = {}
        
        if debug:
            for herb in herbs:
                score = random.randint(1, 5)
                ranked_herbs[herb] = score
        
        else:
            ranked_herbs = pytrend_score_group(herbs)
            
        #print(ranked_herbs)
        ranked_groups.append(ranked_herbs)
    return ranked_groups

The code first creates a test list of herbs and then breaks it into groups of 5 using the break_into_groups function.Then, it ranks each group of herbs by popularity using the rank_by_popularity function, which can use either a random score or Google Trends data depending on the debug parameter. The debug parameter is set to False, so Google Trends data will be used. The final output will be a list of dictionaries where each dictionary contains the ranking scores for each herb in a group.


#test list of herbs and then breaks it into groups of 5 
test_herbs = ["St. John's Wort",
'Cranberry',
'Agrimony',
'Cherry stalks',
'Cannabis flowering tops',
'Herbal tea combinations for use in cough and cold',
"St. Benedict's thistle (also blessed thistle, holy thistle or spotted thistle)",
'Lime flower',
'Guarana',
'Rosemary Oil',
'Rosemary leaf',
'Sweet Fennel']

print("Phase 1 - Break into groups:")
print(break_into_groups(test_herbs,times=5))

print("Phase 2 - Rank by popularity:")
rank_by_popularity(break_into_groups(test_herbs,times=5), debug=False)

import json

# Load the JSON file
with open('remedies.json', 'r') as f:
    remedies = json.load(f)



 #Filter remedies by vernacularNames/VERNAC
enriched_remedies = [r for r in remedies if r.get('vernacularNames', {}).get('VERNAC')]

# Filter remedies by commonNames/CNAME and take every 6th item
filtered_remedies = [r for i, r in enumerate(remedies) if r.get('commonNames', {}).get('CNAME') and i % 6 == 0]

# Remove duplicates
unique_remedies = [dict(t) for t in {tuple(r.items()) for r in filtered_remedies}]

# Save the filtered remedies as a new JSON file
with open('filtered_remedies.json', 'w') as f:
    json.dump(unique_remedies, f)


# Filter remedies by commonNames/CNAME and take every 6th item
filtered_remedies = [r for i, r in enumerate(remedies) if r.get('commonNames', {}).get('CNAME') and i % 6 == 0]


#Remove duplicates
unique_remedies = [dict(t) for t in {tuple(r.items()) for r in filtered_remedies}]



# Save the filtered remedies as a new JSON file
with open('filtered_remedies.json', 'w') as f:
    json.dump(unique_remedies, f)
