==> ds/notebooks/02_b_popular_herbs.ipynb <==
### Trying to use a ranking system to determine the __ relative popularity _of different herbs. ###

#!pip install matplotlib tqdm

from pytrends.request import TrendReq
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from tqdm import tqdm
from itertools import islice

**"Herbal Medicines" EMA Dataset**

The code splits the list of herbs into groups of 5.

# Split the list of herbs into sublists of 5 herbs each
herbs=[
"St. John's Wort",
'Cranberry',
'Agrimony',
'Cherry stalks',
'Cannabis flowering tops',
'Herbal tea combinations for use in cough and cold',
"St. Benedict's thistle (also blessed thistle, holy thistle or spotted thistle)",
'Lime flower',
'Guarana',
'Rosemary Oil',
'Rosemary leaf',
'Sweet Fennel',
'Bitter Fennel',
'California poppy',
'Herbal tea combinations for use in gastrointestinal complaints',
'Willow herb',
'Walnut leaf',
'Javanese turmeric',
'Tribulus herb ',
'Couch grass rhizome',
'Green tea',
'Centella',
'Bitter Fennel Fruit Oil',
'Juniper cone berry',
'Fenugreek',
'Dittany of Crete herb',
'White horehound',
'Black Cohosh',
"St. John's Wort",
'Liquorice Root',
'Iceland moss',
 "Brewer's yeast ",
 "Salvia miltiorrhiza root and rhizome ('Danshen')",
 'Dandelion root',
 'Rose flower',
 'Castor oil',
 'Senna leaf',
 'Cascara',
 'Hedge mustard',
 'Comfrey Root',
 'Ironwort',
 'Senna pods',
 'Rhubarb',
 'Dried juice of Aloes leaves',
 'Frangula bark',
 'Nettle root',
 'Arctic root',
 'Nettle Leaf',
 'Chicory root',
 'Bladderwrack',
 'Pumpkin seed',
 'Ash Leaf',
 'Eucalyptus leaf',
 'Gumweed herb',
 'Restharrow root',
 'Tea-tree oil',
 'Dried Bilberry fruit',
 'Pygeum africanum bark',
 'Majoram',
 'Mouse-ear hawkweed',
 'Fresh Bilberry fruit',
 'Knotgrass herb',
 'Matricaria Flower',
 'Sandy everlasting',
 'Ginkgo leaf',
 'Java Tea',
 'Mastic (Mastix, Pistaciae lentisci resina)',
 'Maté Leaf',
 'Bergamot oil',
 'Herbal tea combinations for use as sedatives',
 'Herbal tea combinations for use in loss of appetite',
 'Wild Pansy',
 'Saw Palmetto Fruit',
 "Butcher's Broom",
 'Clove',
 'Clove oil',
 'Three-lobed sage leaf',
 'Milkthistle Fruit',
 'Evening primrose oil',
 'Mallow leaf',
 'Kava-kava',
 'Lovage root',
 'Peony root, red',
 'Ginseng',
 'Peony root, white',
 'Lemon verbena leaf',
 'Pink rock-rose',
 'Bogbean leaf',
 'Cola',
 'Garlic',
 'Cinnamon Bark Oil',
 'Raspberry leaf',
 'Cinnamon',
 'peppermint leaf',
 'Yarrow',
 'Eucalyptus Oil',
 'Horse-chestnut bark',
 'Lavender',
 'Arnica flower',
 'Woody nightshade stem',
 'Roman Chamomile Flower',
 'Caraway oil',
 'European Goldenrod',
 'Juniper Oil',
 'Feverfew',
 'Thyme oil',
 'Rupturewort',
 'Meadowsweet',
 'Meadowsweet Flower',
 'Horse-Chestnut Seed',
 'Peppermint oil',
 'Tormentil',
 'Oak Bark',
 'Wormwood',
 'Dandelion Root with Herb',
 'Burdock Root',
 'Caraway fruit',
 'Valerian Root and Hop Strobile',
 'Dandelion Leaf',
 'Pelargonium root',
 'Ginger',
 'Caraway fruit',
 'Hamamelis Leaf',
 'Hamamelis Bark',
 'Hamamelis Distillate',
 'Malabar-nut leaf',
 'Oat Herb',
 'Oat Fruit',
 'Shepherds Purse',
 'Polypody Rhizome',
 'Yarrow Flower',
 'Wild Strawberry Leaf',
 'Motherwort',
 'Gentian Root',
 'Green bean pod',
 'Mallow flower',
 'Narrow-leaved coneflower root',
 'Artichoke Leaf',
 'Turmeric',
 'Agnus Castus Fruit',
 'Pale Coneflower Root',
 'Ivy leaf',
 'Soya',
 'Black Cohosh',
 'Elder Flower',
 'Mullein Flower',
 'Calendula Flower',
 'Angelica sinensis root',
 'Kalmegh',
 'Winter-cherry root',
 'Melilot',
 'Bearberry Leaf',
 'Blackcurrant Leaf',
 'Soya-bean lecithin',
 'Grapevine Leaf',
 'Olive leaf',
 'Purple Coneflower Root',
 'Willow Bark',
 'Diuretic herbal tea combinations',
 'Sage Leaf',
 'Soya-bean oil',
 'Boldo Leaf',
 'Devil’s Claw Root',
 'Marshmallow Root',
 'Thyme',
 'Hop strobile',
 'Anise Oil',
 'Aniseed',
 'Primula flower',
 'Passion Flower',
 'Peru balsam',
 'Thyme and Primula root',
 'Valerian essential oil',
 'Hawthorn Leaf and Flower',
 'Horsetail herb',
 'Valerian Root',
 'Matricaria Oil',
 'Purple Coneflower Herb',
 'Eleutherococcus',
 'Centaury',
 "Cat's Claw",
 'Birch Leaf',
 'Capsicum',
 'Katula',
 'Melissa leaf',
 'Lavender Oil',
 'Ispaghula husk',
 'Psyllium seed',
 'Ispaghula Seed',
 'Elderberry',
 'Marigold',
 'Mistletoe',
 'Primula root',
 'Silver lime flower',
 'Onion',
 'Ribwort Plantain',
 'Greater Celandine',
 'Fumitory',
 'Eyebright',
 'Nettle Herb',
 'Sage Oil',
 'Linseed']


len(herbs)

**Groups of five herbs**

then ranks them by popularity using a scoring system. The herbs are then split into groups for the next round based on their popularity scores.

def break_into_groups(my_list, times):
    sublists = []
    for i in range(0, len(my_list), times):
        sublists.append(my_list[i:i+5])
    return sublists

groups = break_into_groups(herbs, 5)
groups

len(groups)

**Random ranked groups**

**Trying to use a ranking system to determine the relative popularity of different herbs.**

Rank them by popularity using a scoring system.
The herbs are then split into groups for the next round based on their popularity scores. This ranking system ranks herbs by popularity in groups and assigns points to the herbs based on their popularity. Then I create a dataframe from the herb scores, sorts it by the score and select the column with the herb names. 

Finally, we break the herbs into new groups based on the sorted order of the herb names.

import random
from typing import List, Dict

def rank_by_popularity(group: List[List[str]]) -> List[Dict[str, int]]:
    ranked_groups = []
    for herbs in group:
        ranked_herbs = {}
        for herb in herbs:
            score = random.randint(1, 5)
            ranked_herbs[herb] = score
        ranked_groups.append(ranked_herbs)
    return ranked_groups
    
ranked_groups = rank_by_popularity(groups)
print(ranked_groups)

import random
from typing import List, Dict

def rank_by_popularity(group: List[List[str]]) -> List[Dict[str, int]]:
    ranked_groups = []
    for herbs in group:
        ranked_herbs = {}
        for herb in herbs:
            score = random.randint(1, 5)
            ranked_herbs[herb] = score
        ranked_groups.append(ranked_herbs)
    return ranked_groups
    
ranked_groups = rank_by_popularity(groups)
print(ranked_groups)

 **Tqdm module** 
 Display a progress bar while iterating through 10 rounds of processing. Within each round, the code ranks the herbs in each input group using the rank_by_popularity function defined in the previous code block. The code then updates a dictionary called herb_score with the scores for each herb.

from tqdm import tqdm
import time

herb_score = {herbs: 0 for herbs in herbs} #dict with 203 keys and all zeroes
ROUNDS = 10

for rnd in tqdm(range(ROUNDS)):
    
    # we need to iterate the groups
    for group in groups:

        #rank the herbs inside the group
        ranked_group = rank_by_popularity([group])
        #assign the points of the herb on herb_score
        ranked_group = ranked_group[0] #fix to transform ranked_group as a dictionary
        
        for herb, score in ranked_group.items():
            herb_score[herb] += ranked_group[herb]
            
        # create a dataframe from herb_score
    df = pd.DataFrame.from_dict(herb_score, orient='index', columns=['score'])
    df.index.name = 'herb'
    df.reset_index(inplace=True)
    time.sleep(1)

# sort the dataframe by score and herb
df = df.sort_values(by=['score','herb'], ascending=False)

# get the sorted herbs
sorted_herbs = df['herb'].tolist()

# break the sorted herbs into groups
groups = break_into_groups(sorted_herbs,times=len(groups))

# re-initialize herb_score to a dictionary with 203 keys and all zeroes
herb_score = {k: 0 for k in herbs}
          
#sort the dataframe by score and herb
df = df.sort_values(by=['score', 'herb'], ascending=False)

herb_score

ranked_groups

**PyTrends**

1. Approach with pytrend that didnt go work:

The issue with the code is that it does not handle the TooManyRequestsError, which can occur when making too many requests to the Google Trends API in a short amount of time. 

To handle this error I tried  to use a try-except block to wait for 5 seconds before making the next request. 

#from pytrends.exceptions import ResponseError, TimeoutError
from pytrends.exceptions import TooManyRequestsError
from tqdm import tqdm
import time

def get_popularity(herb, pytrend):
    # Wait for 5 seconds before making the request
    time.sleep(5)
    try:
        pytrend.build_payload(kw_list=[herb])
        interest_over_time_df = pytrend.interest_over_time()
        # Return the average interest over time for the given herb
        return interest_over_time_df.mean()[0]
    except (TooManyRequestsError):
        # Wait for 5 seconds before making the next request
        time.sleep(5)
        return get_popularity(herb, pytrend)

# Split the herbs into groups of 5
herbs_split = [herbs[i:i+5] for i in range(0, len(herbs), 5)]

# Loop over each sublist of 5 herbs
for herb_group in tqdm(herbs_split):
    # Convert each herb to a string and join them with commas
    herb_group_str = ",".join(str(herb) for herb in herb_group)

    # Initialize a new pytrends object
    pytrend = TrendReq()

    # Get the popularity for each herb in the group
    herb_popularity = {}
    for herb in herb_group:
        herb_popularity[herb] = get_popularity(herb, pytrend)

    # Rank the herbs by popularity
    ranked_herbs = sorted(herb_popularity, key=herb_popularity.get, reverse=True)

    # Split the ranked herbs into groups of 7
    herb_groups = [ranked_herbs[i:i+7] for i in range(0, len(ranked_herbs), 7)]

    # Create a list of groups for the current round
    groups = []
    for group in herb_groups:
        groups.append(list(group))

    # Initialize a dictionary with each herb's score set to 0
    herb_score = {herb: 0 for herb in herbs}

    # Iterate over each group and assign points to each herb in the group
    for group in groups:
        for herb in group:
            herb_score[herb] += len(group) - group.index(herb)

    # Create a dataframe from the herb scores and sort it by score
    df = pd.DataFrame(list(herb_score.items()), columns=['herb', 'score'])
    df = df.sort_values(by=['score', 'herb'], ascending=False)

    # Get the sorted herbs
    sorted_herbs = df['herb'].tolist()

    # Break the sorted herbs into groups for the next round
    groups = break_into_groups(sorted_herbs)
