==> ds/notebooks/02_b_remedies_pytrends.ipynb <==
import os
#os.chdir('pytrends/')

import sys
sys.path.insert(0,os.getcwd())

print(os.getcwd())
print(sys.path)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import random
from typing import List, Dict

from pytrends.request import TrendReq

from tqdm import tqdm
from itertools import islice
from IPython.display import display
import time

from warnings import warn
import json

def pytrend_score_group(terms : list, print_table : bool = False) -> Dict[str,int]:
    pt = TrendReq(retries=5)
    pt.build_payload(terms)
    tmp_df = pt.interest_over_time()
    tmp_df = tmp_df[terms].mean()
    
    if print_table:
        display(tmp_df)
        
    ranks = tmp_df.rank().to_dict()
    return ranks 


def break_into_groups(lst, group_size):
    """
    Breaks a list into groups of a specified size.
    
    Args:
        lst (list): The list to be broken into groups.
        group_size (int): The size of each group.
    
    Returns:
        list: A list of lists, where each inner list contains `group_size`
        elements from the original list.
    """
    return [lst[i:i+group_size] for i in range(0, len(lst), group_size)]


# Load the JSON file
with open('remedies.json', 'r') as f:
    remedies = json.load(f)

remedies[:1]

# Filter remedies by VERNAC
enriched_remedies = [remedy for remedy in remedies if  remedy['VERNAC']]

# Print the number of enriched remedies
print(f"Number of enriched remedies: {len(enriched_remedies)}")

 #Filter enriched_remedies by CNAME
filtered_remedies = [remedy for remedy in enriched_remedies if remedy.get('CNAME')]
print(f"Number of filtered remedies: {len(filtered_remedies)}")

remedies = list({r['TAXON']: r for r in filtered_remedies}.values())
print(f"Number of unique remedies: {len(remedies)}")

remedies_names = [remedy['TAXON'][0] if isinstance(remedy['TAXON'], list) else remedy['TAXON'] for remedy in remedies if remedy.get('TAXON')]
print(remedies_names)

# Break remedies into groups of 5
groups = break_into_groups(remedies_names, 5)

groups
