==> ds/notebooks/02_c_data_enrichment.ipynb <==
import re
import pandas as pd
import numpy as np
import PyPDF2

### NaturDoc - TL BL WT 22-23

# Data enrichment:

## Imports:

### Duke:

mod_duke_df = pd.read_csv("../data/DUKE/PIVOT_ETHNOBOT.csv")

### KEW:

##### _Note: reduced and pivoted WCVP datasets:_

As mentioned in the data_transformation notebook, some of the csv files were too large to be uploaded to Github. By creating a copy of wcvp_distribution.csv and wcvp_names.csv containing only columns that were useful to us and by pivoting the entire table, the number of rows and columns was reduced significantly for each file.

dis_df = pd.read_csv("../data/WCVP/STR_REDUCED_PIVOT_wcvp_distribution.csv", sep=",")
names_df = pd.read_csv("../data/WCVP/REDUCED_wcvp_names.csv", sep=",")

### WHO Monographs:

who_01_pdf = open('../data/WHO/monograph_01.pdf', 'rb')

who_01_reader = PyPDF2.PdfReader(who_01_pdf)

who_01_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_spacequotesingle = re.compile(" ’")
pattern_spacequotedouble = re.compile(" ”")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_01_reader.pages)):
    page_obj = who_01_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_spacequotesingle, "’", text)
    text = re.sub(pattern_spacequotedouble, "”", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_01_pages.append(lines)
    
who_01_pdf.close()

p_01_50 = who_01_pages[7 : 57]
p_51_58 = who_01_pages[75 : 83]
p_59_60 = who_01_pages[57 : 59]
p_61_66 = who_01_pages[69 : 75]
p_67_76 = who_01_pages[59 : 69]
p_77_end = who_01_pages[83 : ]

who_01_pages_corrected = p_01_50 + p_51_58 + p_59_60 + p_61_66 + p_67_76 + p_77_end

## Enriching the Duke dataset:

duke_names = mod_duke_df["TAXON"].unique().tolist()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")

# Definitions with in-keyword:

who_01_def = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Definition" in line:
            n_lines = 1
            while page[i + n_lines] != "Synonyms" and i + n_lines < max_len:
                n_lines += 1
            output = " ".join(page[i + 1 : i + n_lines])

# adding Matricaria chamomilla to Chamomilla recutita's definition to create a match with Duke later
            if "Chamomilla recutita" in output:
                output = "Matricaria chamomilla" + output
            who_01_def.append(output)

# Vernacular names with in-keyword:

who_01_vernacular = list()
reg_pattern = re.compile(r"\.")
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Selected vernacular names" in line:
            n_lines = 1
            while page[i + n_lines] != "Description" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])
            
            output =  output.replace(".", ",")
            
            # Sometimes taxonomic names appear in the WHO document names as a header of sorts:
            for name in duke_names:
                output = output.replace(name, ",")
            
            output =  output.replace(", ", ",")

            who_01_vernacular.append(output)

# Clinical medicine with in-keyword:

who_01_clinical = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" in line:
            n_lines = 1
            while "Uses described in pharmacopoeias and in traditional" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            # in case the description gets cut off, drop the incomplete sentence:
            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."
            
            who_01_clinical.append(output)

# Traditional medicine with in-keyword:

who_01_traditional = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in pharmacopoeias and in traditional" in line:
            n_lines = 2
            while "Uses described in folk medicine" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_traditional.append(output)

# Folk with in-keyword:

who_01_folk = list()
  
for page in who_01_pages_corrected[3:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in folk medicine" in line:
            n_lines = 2
            while page[i + n_lines] != "Pharmacology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 2 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 2 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_folk.append(output)

who_01_con = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Contraindications" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Warnings" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_con.append(output)
            
who_01_warn = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Warnings" in line and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "Precautions" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_warn.append(output)
            
who_01_adv = list()

ref_pattern = re.compile(r"[0-9]+\.")  

for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Adverse reactions" in line and "see" not in line.lower() and not re.search(ref_pattern, line):
            n_lines = 1
            while page[i + n_lines] != "Posology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_adv.append(output)

who_01_pos = list()
  
for page in who_01_pages_corrected[4:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Posology" in line and "and" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "References" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                # output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                # output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_01_pos.append(output)

Replace WHO null values with empty string:

null_values = ["No information available.", "None.", "."]

def make_nan(target_list: list):
    for i, value in enumerate(target_list):
        if value in null_values:
            target_list[i] = ""

make_nan(who_01_clinical)
make_nan(who_01_traditional)
make_nan(who_01_folk)
make_nan(who_01_con)
make_nan(who_01_warn)
make_nan(who_01_adv)
make_nan(who_01_pos)

who_01_remedies = list()

for i, definition in enumerate(who_01_def):

    for name in duke_names:

        remedy_dict = dict()
        name = name.split(" ")

        # some taxonomic names have an "x", indicating hybrids
        if "x" in name: 
            name.remove("x")

        if name[0] in definition and name[1] in definition:
            remedy_dict["NAME"] = " ".join(name)
            remedy_dict["DEF"] = definition
            remedy_dict["VERNAC"] = who_01_vernacular[i]
            remedy_dict["CLINICAL"] = who_01_clinical[i]
            remedy_dict["TRADITIONAL"] = who_01_traditional[i] 
            remedy_dict["FOLK"] = who_01_folk[i]
            remedy_dict["CONTRAINDICATION"] = who_01_con[i]
            remedy_dict["WARNING"] = who_01_warn[i]
            remedy_dict["ADVERSE"] = who_01_adv[i]
            remedy_dict["POSOLOGY"] = who_01_pos[i]

        if remedy_dict:
            who_01_remedies.append(remedy_dict)

### Preparing the Duke dataframe:

When reading the modified ETHNOBOT dataset, a new serial index is created, so we need to drop the "Unnamed: 0" column.

mod_duke_df.columns

target_cols = ["TAXON", "ACTIVITY", "CNAME", "FAMILY", "GENUS", "SPECIES"]
rich_duke_df = mod_duke_df[target_cols]
rich_duke_df.head()

### Adding data from WHO monographs to the reshaped dataset:

Preparing new columns in the enriched dataframe for the information stored in the remedy dictionary. Since we previously replaced all np.NaN values with a string "nan", the default value will be the same here:

rich_duke_df["VERNAC"] = "nan"
rich_duke_df["CLINICAL"] = "nan"
rich_duke_df["TRADITIONAL"] = "nan"
rich_duke_df["FOLK"] = "nan"
rich_duke_df["CONTRAINDICATION"] = "nan"
rich_duke_df["WARNING"] = "nan"
rich_duke_df["ADVERSE"] = "nan"
rich_duke_df["POSOLOGY"] = "nan"

rich_duke_df.head()

Concatenating the dictionary values to the correct dataframe column:

def store_dict_values(remedy: dict, key: str):
    name_filt = (rich_duke_df["TAXON"] == remedy["NAME"])
    if  "nan" in rich_duke_df.loc[name_filt,[key]].values:
        rich_duke_df.loc[name_filt,[key]] = remedy[key]
    else:
        rich_duke_df.loc[name_filt,[key]] = rich_duke_df.loc[name_filt,[key]] + remedy[key]

skip_keys = ["NAME", "DEF"]

for i, herb in enumerate(who_01_remedies):
    for key in herb.keys():
        if key in skip_keys:
            continue
        store_dict_values(herb, key)

To better gauge the success of the previous code, we filter the dataframe to target one of the herbs that we know was in the monograph:

filt_alliumsativum = (rich_duke_df["TAXON"] == "Allium sativum")
rich_duke_df.loc[filt_alliumsativum]

Verifying that chamomille is also added to the dataset now:

filt_chamomilla = (rich_duke_df["TAXON"] == "Matricaria chamomilla")
rich_duke_df.loc[filt_chamomilla]

As we can see, this code successfully added the dictionary values to the dataframe.

### Opening, loading and reading the remaining monographs:

Since every monograph necessitates a lot of modifications of the extraction algorithm, we will only extract the data from the first two monographs out of the four available. As a proof of concept, we think this suffices and already provide in-depth data for up to 70 herbs in the Duke dataset.

# 2:

who_02_pdf = open('../data/WHO/monograph_02.pdf', 'rb')

who_02_reader = PyPDF2.PdfReader(who_02_pdf)

who_02_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_02_reader.pages)):
    page_obj = who_02_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_02_pages.append(lines)
    
who_02_pdf.close()

# 3:

who_03_pdf = open('../data/WHO/monograph_03.pdf', 'rb')

who_03_reader = PyPDF2.PdfReader(who_03_pdf)

who_03_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_03_reader.pages)):
    page_obj = who_03_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_03_pages.append(lines)
    
who_03_pdf.close()

# 4:

who_04_pdf = open('../data/WHO/monograph_04.pdf', 'rb')

who_04_reader = PyPDF2.PdfReader(who_04_pdf)

who_04_pages = list()

pattern_fi = re.compile("ﬁ")
pattern_fl = re.compile("ﬂ")
pattern_minus = re.compile("–")
pattern_brackets = re.compile("\([\s0-9,-]+\)")
pattern_spacedot = re.compile(" \.")
pattern_spacecomma = re.compile(" ,")
pattern_doublespace = re.compile("\s\s+")
pattern_dashspace = re.compile("-\s+")
  
for page in range(len(who_04_reader.pages)):
    page_obj = who_04_reader.pages[page]
    text = page_obj.extract_text()
    text = re.sub(pattern_fi, "fi", text)
    text = re.sub(pattern_fl, "fl", text)
    text = re.sub(pattern_minus, "-", text)
    text = re.sub(pattern_brackets, "", text)
    text = re.sub(pattern_spacedot, ".", text)
    text = re.sub(pattern_spacecomma, ",", text)
    text = re.sub(pattern_doublespace, " ", text)
    text = re.sub(pattern_dashspace, "", text)
    lines = text.split("\n")
    who_04_pages.append(lines)
    
who_04_pdf.close()

pattern_formula = re.compile(r"\.[A-Za-z0-9]+.*")

### Extracting data from the second WHO monograph:

# 2:

# Definitions with in-keyword:

who_02_def = list()
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Definition" in line:
            n_lines = 1
            while page[i + n_lines] != "Synonyms" and i + n_lines < max_len:
                n_lines += 1
            
            output = " ".join(page[i + 1 : i + n_lines])

# adding Matricaria chamomilla to Chamomilla recutita's definition to create a match with Duke later
            if "Tanacetum parthenium" in output:
                output = "Chrysanthemum parthenium" + output
            who_02_def.append(output)

# Vernacular names with in-keyword:

who_02_vernacular = list()
reg_pattern = re.compile(r"\.")
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Selected vernacular names" in line:
            n_lines = 1
            while page[i + n_lines] != "Description" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
            else:
                output = " ".join(page[i + 1 : i + n_lines])
            
            output =  output.replace(".", "")
            
            # Sometimes taxonomic names appear in the WHO document names as a header of sorts:
            for name in duke_names:
                output = output.replace(name, ",")
            
            who_02_vernacular.append(output)

# Clinical medicine with in-keyword:

who_02_clinical = list()
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses supported by clinical data" in line:
            n_lines = 1
            while "Uses described in pharmacopoeias and in traditional" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            # in case the description gets cut off, drop the incomplete sentence:
            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."
            
            who_02_clinical.append(output)

# Traditional medicine with in-keyword:

who_02_traditional = list()
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in pharmacopoeias" in line or "Uses reported in pharmacopoeias" in line:
            n_lines = 1

            while "Uses described in folk medicine" not in page[i + n_lines] and i + n_lines < max_len:
                n_lines += 1

            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if not output:
                print(line)
                print(page[i + 1])
                print(page[i + n_lines])
                print(page[i + n_lines + 1])

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_traditional.append(output)

# Folk with in-keyword:

who_02_folk = list()
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Uses described in folk medicine" in line:
            n_lines = 1
            while page[i + n_lines] != "Pharmacology" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_folk.append(output)
            
who_02_con = list()

for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Contraindications" in line and ".)" not in line.lower() and "and" not in line.lower() and "see" not in line.lower():
            n_lines = 1
            
            while page[i + n_lines] != "Warnings" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if ". Warnings" in output:
                output = "".join(output.split(". Warnings")[0])
                output = output + "."

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_con.append(output)

            
who_02_warn = list()

warning_exception = "androgen and estrogen metabolism. Warnings"
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Warnings" in line and "and" not in line.lower() and "see" not in line.lower() or warning_exception in line:
            n_lines = 1
            while page[i + n_lines] != "Precautions" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_warn.append(output)
            
who_02_adv = list()

ref_pattern = re.compile(r"[0-9]+\.")  

for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Adverse reactions" in line and "in" not in line.lower() and "and" not in line.lower() and "see" not in line.lower() and not re.search(ref_pattern, line):
            n_lines = 1
            while page[i + n_lines] != "Dosage forms" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_adv.append(output)

who_02_pos = list()
  
for page in who_02_pages[10:]:
    max_len = len(page) - 1
    for i, line in enumerate(page):
        if "Posology" in line and "and" not in line.lower() and "see" not in line.lower():
            n_lines = 1
            while page[i + n_lines] != "References" and i + n_lines < max_len:
                n_lines += 1
            if i + n_lines == max_len:
                output = " ".join(page[i + 1 : i + n_lines + 1])
                # output = re.sub(pattern_formula, ".", output)
            else:
                output = " ".join(page[i + 1 : i + n_lines])
                # output = re.sub(pattern_formula, ".", output)

            if "." != output[-1]:
                output = output.split(".")
                output = output[:-1]
                output = ".".join(output)
                output = output + "."

            who_02_pos.append(output)

#### Remove nan values:

make_nan(who_02_clinical)
make_nan(who_02_traditional)
make_nan(who_02_folk)
make_nan(who_02_con)
make_nan(who_02_warn)
make_nan(who_02_adv)
make_nan(who_02_pos)

#### Verify lengths:

print(len(who_02_def))
print(len(who_02_clinical))
print(len(who_02_traditional))
print(len(who_02_folk))
print(len(who_02_con))
print(len(who_02_warn))
print(len(who_02_adv))
print(len(who_02_pos))

### Generating the dictionary:

who_02_remedies = list()

for i, definition in enumerate(who_02_def):

    for name in duke_names:

        remedy_dict = dict()
        name = name.split(" ")

        # some taxonomic names have an "x", indicating hybrids
        if "x" in name: 
            name.remove("x")

        if name[0] in definition and name[1] in definition:
            remedy_dict["NAME"] = " ".join(name)
            remedy_dict["DEF"] = definition
            remedy_dict["VERNAC"] = who_02_vernacular[i]
            remedy_dict["CLINICAL"] = who_02_clinical[i]
            remedy_dict["TRADITIONAL"] = who_02_traditional[i] 
            remedy_dict["FOLK"] = who_02_folk[i]
            remedy_dict["CONTRAINDICATION"] = who_02_con[i]
            remedy_dict["WARNING"] = who_02_warn[i]
            remedy_dict["ADVERSE"] = who_02_adv[i]
            remedy_dict["POSOLOGY"] = who_02_pos[i]

        if remedy_dict:
            who_02_remedies.append(remedy_dict)

#### Checking for missing herbs:

matches = list()

for definition in who_02_def:
    for herb in who_02_remedies:
        if definition in herb["DEF"]:
            matches.append(definition)

rest = list()

for definition in who_02_def:
    if definition not in matches:
        rest.append(definition)

print("Length of matches:", len(matches), "\nLength of rest:", len(rest), end="\n\n")
print(*rest, sep="\n")

Checking the Duke dataset for matches of other synonyms of the plants, we have found _Tanacetum parthenium_ listed under its synonym _Chrysanthemum parthenium_:

filt_tanacetumparthenium = (rich_duke_df["TAXON"].str.contains("parthenium"))
rich_duke_df.loc[filt_tanacetumparthenium]

A code similar to the one that ensured a match for chamomile was already included.

### Adding data from WHO monographs to the reshaped dataset:

skip_keys = ["NAME", "DEF"]

for i, herb in enumerate(who_02_remedies):
    for key in herb.keys():
        if key in skip_keys:
            continue
        store_dict_values(herb, key)

Verifying one of the added herbs:

filt_chrysanthemumparthenium = (rich_duke_df["TAXON"] == "Chrysanthemum parthenium")
rich_duke_df.loc[filt_chrysanthemumparthenium]

#### NAN values:

Replace string nan and empty values with proper np.nan:

nan_toreplace = ["nan", ""]
rich_duke_df = rich_duke_df.replace(nan_toreplace, np.nan)

filt_chrysanthemumparthenium = (rich_duke_df["TAXON"] == "Chrysanthemum parthenium")
rich_duke_df.loc[filt_chrysanthemumparthenium]

Replace the string nan values created from pivoting the table with an aggregate function:

rich_duke_df.ACTIVITY = rich_duke_df.ACTIVITY.str.replace('nan,', '', regex=True)
rich_duke_df.CNAME = rich_duke_df.CNAME.str.replace(',nan', '', regex=True)
rich_duke_df.ACTIVITY = rich_duke_df.ACTIVITY.str.replace('nan,', '', regex=True)
rich_duke_df.CNAME = rich_duke_df.CNAME.str.replace(',nan', '', regex=True)

rich_duke_df.ACTIVITY

rich_duke_df

#### Checking the enriched entries:

temp_df = rich_duke_df.dropna(subset=["VERNAC"])
temp_df.loc[:, ["TAXON", "ACTIVITY", "CNAME", "VERNAC", "CLINICAL", "TRADITIONAL", "FOLK", "CONTRAINDICATION", "WARNING", "ADVERSE", "POSOLOGY"]]

## Export dataframe to json:

Exporting to .json file (different formats, see https://datatofish.com/export-pandas-dataframe-json/):

rich_duke_df.to_json(r"../output/remedies.json", orient = "records")

## Updating to newest naming convention for the dataset:

### Remedy dataset:

Importing the .json file so that we do not have to rerun the entire notebook code again (also for future adjustments).

rich_duke_df = pd.read_json("../output/remedies.json")
rich_duke_df.head()

Creating a ratings column containign empty lists (the apply will move the column to the first position, so we are running this code before renaming the other columns so that it will appear at the end, as it should):

rich_duke_df['ratings'] = ''
rich_duke_df['ratings'] = rich_duke_df['ratings'].apply(list)

col_old = rich_duke_df.columns.tolist()
col_new = ['remedyName', 
           'medicinalUses', 
           'commonNames',
           'family',
           'genus',
           'species',
           'vernacularNames',
           'treatmentClinical',
           'treatmentTraditional',
           'treatmentFolk',
           'contraindication',
           'warnings',
           'adverseEffects',
           'posology']
col_dict = dict(zip(col_old, col_new))
col_dict

for old_col, new_col in col_dict.items():
    rich_duke_df.rename(columns={old_col: new_col}, inplace = True)
print(rich_duke_df.columns.tolist())

#### Adding empty columns for future values:

The database should also store a list of rating as well as the average of these ratings.

rich_duke_df['ratings']

rich_duke_df['ratingAverage'] = 0.0
rich_duke_df['totalNumberofRatings'] = 0

rich_duke_df.head()

### Symptoms dataset:

import pandas as pd

symptoms_df = pd.read_json("../output/symptoms.json")
symptoms_df.head()

symptoms_df.rename(columns={'SYMPTOM': 'symptomName'}, inplace = True)
symptoms_df.head()

### Exporting updated dataframe to json:

rich_duke_df.to_json(r"../output/remedies.json", orient = "records")

symptoms_df.to_json(r"../output/symptoms.json", orient = "records")

test_df = pd.read_json("../output/symptoms.json")
test_df.head(2)

## Looking Ahead:

This notebook concludes our work with the original datasets. The following notebooks will focus on our Machine Learning approach, starting with some general observations and explorations that lay the groundwork for our Machine Learning code.
