==> ds/notebooks/03_b_ml_data_preparation.ipynb <==
### NaturDoc - TL BL WT 22-23

# Data clustering:

## Preparing Activity and Symptom Datasets for Word Embeddings:

Preparing our activities and symptoms data:

import numpy as np
import pandas as pd
from sklearn.datasets import make_blobs

## Data from the dataset:

#### Import duke activities:

duke_enriched_df = pd.read_json("../output/remedies.json")
activities_df = pd.read_json("../output/symptoms.json")

duke_enriched_df.head()

activities_df.shape

activities_df.head()

#### Import other list of symptoms:

gb_symptoms_df = pd.read_csv("../data/GOOGLE_SYMPTOMS/2022_country_weekly_2022_GB_weekly_symptoms_dataset.csv")

df_columns = list(gb_symptoms_df.columns)
print(df_columns[:15])

symptoms = list()

for col in df_columns:
    if "symptom:" not in col:
        continue
    col = col.replace("symptom:", "")
    col = col.capitalize()
    symptoms.append(col)

print(symptoms[::10])

#### Concatenate duke activities and symptoms:

# activities = symptoms_df.values.tolist()
symptoms_df = pd.DataFrame(symptoms)
symptoms_df.rename(columns={0: "symptomName"}, inplace=True)
symptoms_df[["is_symptom", "is_activity"]] = (1, 0)
symptoms_df.head()

activities_df[["is_symptom", "is_activity"]] = (0, 1)
activities_df.head()

Only take the unique values!

print(len(activities_df))
print(len(symptoms_df))
activities_symptoms_df = pd.concat([activities_df, symptoms_df],axis=0)
print("Should be:", len(activities_df) + len(symptoms_df))
print("Is:", len(activities_symptoms_df))
activities_symptoms_df.reset_index(inplace=True)
activities_symptoms_df.drop(labels="index", axis=1, inplace=True)
activities_symptoms_df.head()

However, actually not all unique (combinations are, but some symptomNames were present in both):

activities_symptoms_df.nunique()

activities_symptoms_df["symptomName"]

activities_symptoms_df["symptomName"].duplicated()

duplicate_activities = (activities_symptoms_df["symptomName"].duplicated("last"))
filt_df = activities_symptoms_df[duplicate_activities]
filt_df

duplicate_symptoms = (activities_symptoms_df["symptomName"].duplicated("first"))
filt_df = activities_symptoms_df[duplicate_symptoms]
filt_df

duplicate_indices = filt_df.index
duplicate_indices

#### Adding a 1 to is_symptom for activities that are also symptoms:

activities_symptoms_df.loc[duplicate_activities, ["is_symptom"]] = 1
activities_symptoms_df[duplicate_activities].head()

activities_symptoms_df.iloc[23]

activities_symptoms_df.drop_duplicates(subset="symptomName")

activities_symptoms_df.drop_duplicates(subset="symptomName", inplace=True)

activities_symptoms_df.head()

Verifying results:

filt = (activities_symptoms_df["is_symptom"] == 1) & (activities_symptoms_df["is_activity"] == 1)
filt_df = activities_symptoms_df[filt]
filt_df

Resetting index:

activities_symptoms_df.reset_index(inplace=True)
activities_symptoms_df.drop(labels=["index"], axis=1, inplace=True)
activities_symptoms_df


### Export to csv:

activities_symptoms_df.to_csv("./activities_symptoms_bool.csv")

# activities_symptoms_series = pd.Series(activities_symptoms_series)
# activities_symptoms_series.to_csv("./activities_symptoms.csv")

## Looking Ahead:

The next notebook will experiment with different word embedding models and generate embedding vectors that will then be used for our clustering and distance matrix.
