==> ds/notebooks/03_c_ml_Word_embeddings-herbs.ipynb <==
**Word embeddings**


Word embeddings are a type of NLP technique that maps words or phrases to vectors in a high-dimensional space, where words with similar meanings are represented by vectors that are close to each other

pip install -U sentence-transformers

from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np



Imports the SentenceTransformer library and two pre-trained models,'all-MiniLM-L6-v2' and 'average_word_embeddings_glove.840B.300d', which are used to convert sentences into high-dimensional vectors, also known as word embeddings.



model1 = SentenceTransformer('all-MiniLM-L6-v2')
model2 = SentenceTransformer('average_word_embeddings_glove.840B.300d')

data = pd.read_csv('activities_symptoms_bool.csv', sep= ",")


# Assign the 'symptomName' column to a variable
symptoms = data['symptomName']


This code block defines the dimensionality of the embedding vectors as 768 
defines two functions that use two different pre-trained models, model1 and model2,
convert a given sentence to an embedding vector. 

# Define the dimensionality of the embedding vectors
EMBEDDING_DIM = 768

# Define a function to convert a sentence to an embedding vector using model1
def sentence_to_vector1(sentence):
    return model1.encode(sentence)

# Define a function to convert a sentence to an embedding vector using model2
def sentence_to_vector2(sentence):
    return model2.encode(sentence)

# Convert each variable in the 'symptomName' column to an embedding vector using model1
embedding_vectors1 = [sentence_to_vector1(symptom) for symptom in symptoms]

# Convert each variable in the 'symptomName' column to an embedding vector using model2
embedding_vectors2 = [sentence_to_vector2(symptom) for symptom in symptoms]

# Create a dataframe to store the embedding vectors for each variable
embedding_df = pd.DataFrame({'Symptom': symptoms, 'Embedding1': embedding_vectors1, 'Embedding2': embedding_vectors2})



embedding_df

embedding_df.to_csv('word2_embeddings.csv', index=False)
