==> ds/notebooks/03_d_ml_distancematrix.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### NaturDoc - TL BL WT 22-23

# Data clustering:

## Generating Distance Matrices and a Symptom-Activity Matching Dictionary:

## Loading Embeddings Data:


symptoms_embeddings = pd.read_csv("../data/embeddings/word_embeddings_dataframe.csv")

The dataframe contains three columns: the symptom name (from the Duke dataset and the google symptom data), and one column for the two embedding models each: the first being `all-MiniLM-L6-v2` and the second being `average_word_embeddings_glove.840B.300d`.

print(symptoms_embeddings.shape)
symptoms_embeddings.head(2)

Extracting a dictionary matching index to symptom name:

dict_symptom = symptoms_embeddings["Symptom"].to_dict()
list(dict_symptom.items())[:5]

#### Transforming:

Reading from the csv, it is no longer a proper list but instead a string containing extra characters:

symptoms_embeddings.loc[0, "Embedding1"][:100]

type(symptoms_embeddings.loc[1, "Embedding1"])

test_list_1 = symptoms_embeddings.loc[0, "Embedding1"].replace("\n", "").replace("[", "").replace("]", "").split(" ")
test_list_2 = symptoms_embeddings.loc[0, "Embedding2"].replace("\n", "").replace("[", "").replace("]", "").split(" ")

test_list_1[:5]

Removing all empty strings:

test_list_1 = [x for x in test_list_1 if x]
test_list_2 = [x for x in test_list_2 if x]
test_list_1[:5]

len(test_list_1)

len(test_list_2)

### Creating useable dataframes:

#### Embedding 1 column:

First, transform content of rows from strings to lists:

symptoms_embeddings.loc[:, "Embedding1"].str.replace("[", "", regex=True).replace("]", "", regex=True).str.split(" ")[1][:10]

def listify_df_values(df_series: pd.Series):
    df_series = df_series.str.replace("\n", "", regex=True)
    df_series = df_series.str.replace("[", "", regex=True).replace("]", "", regex=True)
    df_series = df_series.str.split(" ")
    # df_list = df_list.apply(lambda x: x for x in df_list if x)
    return df_series

embeddings1_series = listify_df_values(symptoms_embeddings.loc[:, "Embedding1"])
# symptoms_embeddings.loc[:, "Embedding1"] = symptoms_embeddings.loc[:, "Embedding1"]

embeddings1_series = embeddings1_series.apply(lambda row: [val for val in row if val])

https://stackoverflow.com/questions/67442107/pandas-expand-explode-dataframe-horizontally

embeddings1_df = pd.DataFrame(embeddings1_series)
embeddings1_df.head(5)

embeddings1_df = pd.concat(
    [embeddings1_df[c].apply(pd.Series).add_prefix(c + "_") for c in embeddings1_df], axis=1
)

embeddings1_df.head(5)

Wrong data type:

embeddings1_df.loc[0, "Embedding1_0"]

type(embeddings1_df.loc[0, "Embedding1_0"])

embeddings1_df = embeddings1_df.apply(pd.to_numeric, errors='coerce')

type(embeddings1_df.loc[0, "Embedding1_0"])

#### Embedding 2 column:

Repeat the above steps for the Embedding_2 column:

Creating the basic dataframe:

embeddings2_series = listify_df_values(symptoms_embeddings.loc[:, "Embedding2"])
embeddings2_series = embeddings2_series.apply(lambda row: [val for val in row if val])

embeddings2_df = pd.DataFrame(embeddings2_series)
embeddings2_df.head()

Exploding the lists of values into their own columns so that every cell only contains a single value:

embeddings2_df = pd.concat(
    [embeddings2_df[c].apply(pd.Series).add_prefix(c + "_") for c in embeddings2_df], axis=1
)

embeddings2_df.head()

Converting the cell values to floats:

embeddings2_df = embeddings2_df.apply(pd.to_numeric, errors='coerce')
type(embeddings2_df.loc[0, "Embedding2_0"])

## Creating the Distance Matrix:

# importing the library
from scipy.spatial import distance_matrix

import math

def generate_distance_matrix(df : pd.DataFrame,
                distance_metric : str = "euclidean") -> pd.DataFrame: # 2.5k x 2.5k
    if distance_metric == "manhattan":
        p = 1
    elif distance_metric == "euclidean":
        p = 2
    elif distance_metric == "chebychev":
        p = math.inf
    else:
        p = 2
    dis_matrix = distance_matrix(df.values, df.values, p)
    dis_df = pd.DataFrame(dis_matrix)
    return dis_df


df_dist_1 = generate_distance_matrix(embeddings1_df)
df_dist_1

## Generating the Final Dictionary:

### A Dictionary of All Symptoms and Activities and their Related Terms:

Pseudocode reminder:

    def generate_dict(df_dist : pd.DataFrame,
                    threshold : float) -> Dict[str: List[str]]:

#### A Look at the Current Values:

To get an idea of what kind of threshold make sense, let's have a look at what the values for a single column look like:

df_dist_1.loc[:, 0].max()

df_dist_1.loc[:, 0].min()

df_dist_1.loc[:, 0].mean()

df_dist_1.loc[:, 0].median()

df_dist_1.loc[1:, 0].min()

A look at min and max values across the entire dataframe: 

print(df_dist_1.max())
df_dist_1.max().max()

print(df_dist_1.copy().replace(0.0, 10).min())
df_dist_1.copy().replace(0.0, 10).min().min()

The maximum value in the similarity matrix is about 1.44254, the minimum is 0.0. 0.0 indicates absolute similarity and in this similarity matrix happens when rows and columns refer to the same initial data entry (so at (0,0), (1,1), (2,2) and so on).The mean lies at around 1.23030, with the median a bit lower at about 1.22825.

Looking at the minimum value where similarity is not referenced relevant to the entry itself, (so in column 0, from row 1 onwards), we have about 1.03327 as a minimum value. We should therefore aim to have a threshold above this.

def generate_dict(df_dist: pd.DataFrame,
                threshold: float) -> dict:
    filt = (df_dist[:] > threshold)
    df_filt = df_dist.copy()
    df_filt[filt] = np.nan
    dict_dist = df_filt.to_dict('dict')
    for i, dic in dict_dist.items():
        to_pop = list()
        for key, value in dic.items():
            if np.isnan(value):
                to_pop.append(key)
            # elif value == 0.0:
            #     to_pop.append(key)
        for target_key in to_pop:
            dic.pop(target_key)
        dict_dist[i] = dic
    return dict_dist

dict_dist = generate_dict(df_dist_1, 0.5)

Taking a look at the first three dictionary entries:

print({k: dict_dist[k] for k in list(dict_dist)[:3]})

dict_dist[0]

dict_symptom[0]

def generate_dict_match(dict_dist: dict) -> dict:   
    dict_match = dict()

    for key, value in dict_dist.items():
        for sub_key in value.keys():
            if dict_symptom[key] not in dict_match:
                dict_match[dict_symptom[key]] = [dict_symptom[sub_key]]
            else:
                dict_match[dict_symptom[key]] = [*dict_match.get(dict_symptom[key]), dict_symptom[sub_key]]
                # [*response.get("match_partial"), match_partial[i_partial]]

    return dict_match

##### Threshold 0.5:

Much too restrictive.

dict_dist = generate_dict(df_dist_1, 0.5)
dict_match = generate_dict_match(dict_dist)
list(dict_match.items())[:5]

##### Threshold 0.8:

This threshold already seems to provide really good results, albeit maybe a bit too narrow.

dict_dist = generate_dict(df_dist_1, 0.8)
dict_match = generate_dict_match(dict_dist)
list(dict_match.items())[:5]

##### Threshold 1.0:

A threshold of 1.0 also provides good results, although in this case the scope might be a little bit too broad. This could be a great threshold for a future functionality proposing remedies for related symptoms.

dict_dist = generate_dict(df_dist_1, 1.0)
dict_match = generate_dict_match(dict_dist)

Taking a closer look at "Abdomen" in particular, the threshold does in fact appear to be a bit too open.

print(generate_dict_match(dict_dist)["Abdomen"])

##### Threshold 0.9:

A threshold of 0.9 yields similar results as a 1.0 threshold:

dict_dist = generate_dict(df_dist_1, 0.9)
print(generate_dict_match(dict_dist)["Abdomen"])
print(generate_dict_match(dict_dist)["Cold"])
print(generate_dict_match(dict_dist)["Abdominal pain"])
print(generate_dict_match(dict_dist)["Xerostomia"])

### A Dictionary of Symptoms Only and their Related Terms from the Duke Activities:

For the sake of the Naturedoc web application, our goal is to provide the users with a more intuitive range of symptoms. As previously discussed, the Duke Database's _Activity_ column contains not just symptoms but also other forms of use (such as _tea_, _spice_, _abortifacient_ etc.). Additionally, these terms are often formatted in a way that is different from how most users are expected to input their symptoms: e.g., various aches are always written as _Ache(Stomach)_ and so on.

For now, we will use the list of symptoms extracted from the Google Symptoms Database. Ideally, this set of symptoms should be further modified and enriched - with specific diseases, synonyms etc. for an optimised user experience. 

For the sake of Naturedoc's proof of concept, we will focus on the 422 symptoms as they exist in the google Database:

activities_symptoms_df = pd.read_csv("../output/activities_symptoms_bool.csv")
activities_symptoms_df.drop(columns="Unnamed: 0", inplace=True)
activities_symptoms_df.head()

This dataframe contains all symptoms and activities from both the Google Dataset and Duke's Database. The _is_symptom_ and _is_activity_ columns indicate which of these sources they originate from.

filt_sym = (activities_symptoms_df["is_symptom"] == 1)
filt_sym_df = activities_symptoms_df[filt_sym]
filt_sym_list = filt_sym_df["symptomName"].values.tolist()

Some symptoms were matched to just themselves or other symptoms that only exist in the Google Dataset, but not Duke's Database.

These should be removed.

filt = (activities_symptoms_df["symptomName"] == "Xerostomia")
filt_df = activities_symptoms_df[filt]
filt_df

In general, symptoms originating from the Google Dataset should probably be removed from the dictionary values, as the might cause issues when querying the database.

To eventually exclude not-activities from the dictionary, we also create a list of entries not in the activities:

filt_not_act = (activities_symptoms_df["is_activity"] == 0)
filt_not_act_df = activities_symptoms_df[filt_not_act]
filt_not_act_list = filt_not_act_df["symptomName"].values.tolist()

We generate the dictionary using a threshold of 0.9:

dict_dist_09 = generate_dict(df_dist_1, 0.9)
dict_09 = generate_dict_match(dict_dist_09)

Keys that exist in the list of Google Symptoms are added to a new dictionary, alongside their values:

dict_09_sym = dict()

for key in dict_09.keys():
    if key in filt_sym_list:
        dict_09_sym[key] = dict_09[key]

print(len(dict_09_sym))
list(dict_09_sym.items())[:5]

Generate the dictionary while removing Google Symptoms from the values inside the dictionary:

def create_dict_sym(dict_dist):  
    dict_sym = dict()

    for sym, list_sym in dict_dist.items():
        if sym not in filt_sym_list:
            continue
        for sub_sym in list_sym:
            if sub_sym in filt_not_act_list:
                continue
            if sym not in dict_sym:
                dict_sym[sym] = [sub_sym]
            else:
                dict_sym[sym] = [*dict_sym.get(sym), sub_sym]
    
    return dict_sym

dict_09_sym = create_dict_sym(dict_09)

print(len(dict_09_sym))
list(dict_09_sym.items())[:5]

A few too many hits for cancers and tumors?

print(dict_09_sym["Abdominal pain"])
print(dict_09_sym["Eye pain"])

dict_dist_085 = generate_dict(df_dist_1, 0.85)
dict_085 = generate_dict_match(dict_dist_085)
dict_085_sym = create_dict_sym(dict_085)
print(len(dict_085_sym))
list(dict_085_sym.items())[:5]

This already removes some of the cancer/tumor matches - however, symptoms like "Common cold" also drop out:

print(dict_085_sym["Abdominal pain"])
print(dict_085_sym["Eye pain"])
print("Common cold" in dict_085_sym.keys())

A threshold of .86 looks like an appropriate compromise for now:

dict_dist_086 = generate_dict(df_dist_1, 0.86)
dict_086 = generate_dict_match(dict_dist_086)
dict_086_sym = create_dict_sym(dict_086)
print(len(dict_086_sym))
list(dict_086_sym.items())[:5]

print(dict_086_sym["Abdominal pain"])
print(dict_086_sym["Eye pain"])
print("Common cold" in dict_086_sym.keys())

Create a new df containing the symptoms from the dictionary to export them to a json:

final_syms_df = pd.DataFrame(sorted(dict_086_sym.keys())).rename(columns = {0: "symptomName"})
final_syms_df.head()

# final_syms_df.to_json("../output/symptoms.json", orient="records")

Export the symptom and matching activities as json:

# import json

# with open("../output/symptom_matches.json", "w") as outfile:
#     json.dump(dict_086_sym, outfile)

### Creating the Dictionary for the Second Embedding Model:

The values of the second embedding model are not usable in the current state. Several of the symptoms end up being matched with with a disproportionate amount of other symptoms and activities. As can be seen in the example of a specific row's value counts below, the exact same value is present 830 times.

#### Distance Matrix:

df_dist_2 = generate_distance_matrix(embeddings2_df)

print("Value counts:", df_dist_2.loc[:, 1].value_counts())

This issue was further explored in the notebook __03_d_ml_distancematrix_emb2_bandaid.ipynb__.

## Looking Ahead:

After the next notebook exploring the errors with generating a distance matrix from the second embedding model, we will dive into clusters using the HDBSCAN* library.
