==> ds/notebooks/03_d_ml_distancematrix_emb2_bandaid.ipynb <==
import pandas as pd
import numpy as np

### NaturDoc - TL BL WT 22-23

# Data clustering:

## Dealing With the Second Embedding Column:

### Loading Embeddings Data:


symptoms_embeddings = pd.read_csv("../data/embeddings/word_embeddings_dataframe.csv")

As a reminder: the dataframe contains three columns: the symptom name (from the Duke dataset and the google symptom data), and one column for the two embedding models each: the first being `all-MiniLM-L6-v2` and the second being `average_word_embeddings_glove.840B.300d`.

print(symptoms_embeddings.shape)
symptoms_embeddings.head(2)

Extracting a dictionary matching index to symptom name:

dict_symptom = symptoms_embeddings["Symptom"].to_dict()

#### Transforming:

Reading from the csv, it is no longer a proper list but instead a string containing extra characters:

symptoms_embeddings.loc[0, "Embedding1"][:100]

type(symptoms_embeddings.loc[1, "Embedding1"])

test_list_1 = symptoms_embeddings.loc[0, "Embedding1"].replace("\n", "").replace("[", "").replace("]", "").split(" ")
test_list_2 = symptoms_embeddings.loc[0, "Embedding2"].replace("\n", "").replace("[", "").replace("]", "").split(" ")

test_list_1[:5]

Removing all empty strings:

test_list_1 = [x for x in test_list_1 if x]
test_list_2 = [x for x in test_list_2 if x]
test_list_1[:5]

### Creating useable dataframes:

def listify_df_values(df_series: pd.Series):
    df_series = df_series.str.replace("\n", "", regex=True)
    df_series = df_series.str.replace("[", "", regex=True).replace("]", "", regex=True)
    df_series = df_series.str.split(" ")
    # df_list = df_list.apply(lambda x: x for x in df_list if x)
    return df_series

#### Embedding 2 column:

Creating the basic dataframe:

embeddings2_series = listify_df_values(symptoms_embeddings.loc[:, "Embedding2"])
embeddings2_series = embeddings2_series.apply(lambda row: [val for val in row if val])

embeddings2_df = pd.DataFrame(embeddings2_series)
embeddings2_df.head()

Exploding the lists of values into their own columns so that every cell only contains a single value:

embeddings2_df = pd.concat(
    [embeddings2_df[c].apply(pd.Series).add_prefix(c + "_") for c in embeddings2_df], axis=1
)

embeddings2_df.head()

Converting the cell values to floats:

embeddings2_df = embeddings2_df.apply(pd.to_numeric, errors='coerce')
type(embeddings2_df.loc[0, "Embedding2_0"])

### Creating the Distance Matrix:

# importing the library
from scipy.spatial import distance_matrix

import math

def generate_distance_matrix(df : pd.DataFrame,
                distance_metric : str = "euclidean") -> pd.DataFrame: # 2.5k x 2.5k
    if distance_metric == "manhattan":
        p = 1
    elif distance_metric == "euclidean":
        p = 2
    elif distance_metric == "chebychev":
        p = math.inf
    else:
        p = 2
    dis_matrix = distance_matrix(df.values, df.values, p)
    dis_df = pd.DataFrame(dis_matrix)
    return dis_df


### Creating the Dictionaries:

def generate_dict(df_dist : pd.DataFrame,
                threshold : float) -> dict:
    filt = (df_dist[:] > threshold)
    df_filt = df_dist.copy()
    df_filt[filt] = np.nan
    dict_dist = df_filt.to_dict('dict')
    for i, dic in dict_dist.items():
        to_pop = list()
        for key, value in dic.items():
            if np.isnan(value):
                to_pop.append(key)
            # elif value == 0.0:
            #     to_pop.append(key)
        for target_key in to_pop:
            dic.pop(target_key)
        dict_dist[i] = dic
    return dict_dist

def generate_dict_match(dict_dist: dict) -> dict:   
    dict_match = dict()

    for key, value in dict_dist.items():
        for sub_key in value.keys():
            if dict_symptom[key] not in dict_match:
                dict_match[dict_symptom[key]] = [dict_symptom[sub_key]]
            else:
                dict_match[dict_symptom[key]] = [*dict_match.get(dict_symptom[key]), dict_symptom[sub_key]]
                # [*response.get("match_partial"), match_partial[i_partial]]

    return dict_match

activities_symptoms_df = pd.read_csv("../output/activities_symptoms_bool.csv")
activities_symptoms_df.drop(columns="Unnamed: 0", inplace=True)
activities_symptoms_df.head()

filt_sym = (activities_symptoms_df["is_symptom"] == 1)
filt_sym_df = activities_symptoms_df[filt_sym]
filt_sym_list = filt_sym_df["symptomName"].values.tolist()

filt_not_act = (activities_symptoms_df["is_activity"] == 0)
filt_not_act_df = activities_symptoms_df[filt_not_act]
filt_not_act_list = filt_not_act_df["symptomName"].values.tolist()

def create_dict_sym(dict_dist):  
    dict_sym = dict()

    for sym, list_sym in dict_dist.items():
        if sym not in filt_sym_list:
            continue
        for sub_sym in list_sym:
            if sub_sym in filt_not_act_list:
                continue
            if sym not in dict_sym:
                dict_sym[sym] = [sub_sym]
            else:
                dict_sym[sym] = [*dict_sym.get(sym), sub_sym]
    
    return dict_sym

## Creating the Dictionary for the Second Embedding Model:

#### Distance Matrix:

df_dist_2 = generate_distance_matrix(embeddings2_df)
df_dist_2

Several entries seem to have matched to an absurd amount of activities:

print("Value counts:", df_dist_2.loc[:, 1].value_counts())

print("Value counts:", df_dist_2.loc[:, 6].value_counts())

Trying other distance metrics results in similar results:

df_dist_2_manhattan = generate_distance_matrix(embeddings2_df, 1)

df_dist_2_chebychev = generate_distance_matrix(embeddings2_df, 3)

print("Value counts:", df_dist_2_manhattan.loc[:, 1].value_counts())

print("Value counts:", df_dist_2_chebychev.loc[:, 1].value_counts())

### Final Dictionaries:

#### Trying Different Threshold:

##### Threshold 2.0:

dict_dist_20 = generate_dict(df_dist_2, 2)
dict_20 = generate_dict_match(dict_dist_20)
dict_20_sym = create_dict_sym(dict_20)
print(len(dict_20_sym))
list(dict_20_sym.items())[:5]

print("Abdominal pain" in dict_20_sym.keys())
print("Eye pain" in dict_20_sym.keys())
print("Common cold" in dict_20_sym.keys())

##### Threshold 5.0:

dict_dist_50 = generate_dict(df_dist_2, 5)
dict_50 = generate_dict_match(dict_dist_50)
dict_50_sym = create_dict_sym(dict_50)
print(len(dict_50_sym))
list(dict_50_sym.items())[:5]

print("Abdominal pain" in dict_50_sym.keys())
print("Eye pain" in dict_50_sym.keys())
print("Common cold" in dict_50_sym.keys())

Due to the erroneous values in the distance matrix, some symptoms match an absurd amount of activities:

print(dict_50_sym["Eye pain"])
print(len(dict_50_sym["Common cold"]))
print(dict_50_sym["Common cold"][::20])

print(dict_50_sym["Avoidant personality disorder"][::20])

## A Bandaid Solution (?):

As a bandaid, we tried to modify the initial dictionary generating code. Unproblematic nested dictionaries are added to a new dictionary, and then the problematic ones will be further processed. This did not really work out too well and is a bad approach anyway:

def generate_dict_patch(df_dist : pd.DataFrame,
                threshold : float) -> dict:
    filt = (df_dist[:] > threshold)
    df_filt = df_dist.copy()
    df_filt[filt] = np.nan
    dict_dist = df_filt.to_dict('dict')
    for i, dic in dict_dist.items():
        to_pop = list()
        for key, value in dic.items():
            if np.isnan(value):
                to_pop.append(key)
        for target_key in to_pop:
            dic.pop(target_key)
        dict_dist[i] = dic
    dict_clear = dict()
    for key, dic in dict_dist.items():
        
# first, add unproblematic dictionaries to dict_clear:
        if len(dic) <= 100:
            dict_clear[key] = dic
            continue

# a new dict lists the indices for each value:
        count_dict = dict()
        for sub_key, val_dis in dic.items():
            if val_dis not in count_dict:
                count_dict[val_dis] = [sub_key]
            else:
                count_dict[val_dis] = [*count_dict.get(val_dis), sub_key]

# checking the count_dict: 
# if a certain key has too many indices as its value, it will be skipped:
        for val, val_i in count_dict.items():
            if len(val_i) > 100:
                continue

# in reverse:
# remaining values will be assigned to correct index keys in dict_clear:
            for i in val_i:
                if i not in dict_clear:
                    dict_clear[i] = {i: val}
                else:
                    dict_clear[i][i] = val

# in case self referencing 0.0 value was removed, add it again:
    target_i = list()
    for key, value in dict_clear.items():
        if key not in value:
            target_i.append(key)
    for i in target_i:
        dict_clear[i] = 0.0

    return dict_clear

print(list(dict_dist_50[6].items())[::20])

dict_dist_patch_50 = generate_dict_patch(df_dist_2, 5)

print(list(dict_dist_patch_50.items())[::20])

dict_patch_50 = generate_dict_match(dict_dist_patch_50)
dict_dist_patch_50_sym = create_dict_sym(dict_patch_50)
print(len(dict_dist_patch_50_sym))
print(list(dict_dist_patch_50_sym.items())[::20])

print("Abdominal pain" in dict_dist_patch_50_sym.keys())
print("Eye pain" in dict_dist_patch_50_sym.keys())
print("Common cold" in dict_dist_patch_50_sym.keys())

Changing thresholds somehow results in keys being removed (maybe they slipped into a territory of having too many matches):

dict_dist_patch_80 = generate_dict_patch(df_dist_2, 8)
dict_patch_80 = generate_dict_match(dict_dist_patch_80)
dict_dist_patch_80_sym = create_dict_sym(dict_patch_80)
print(len(dict_dist_patch_80_sym))
print("Abdominal pain" in dict_dist_patch_80_sym.keys())
print("Eye pain" in dict_dist_patch_80_sym.keys())
print("Common cold" in dict_dist_patch_80_sym.keys())

print(list(dict_dist_patch_80_sym.items())[::20])

After more cautious threshold changes, 5.1 seems the best so far:

dict_dist_patch_51 = generate_dict_patch(df_dist_2, 5.1)
dict_patch_51 = generate_dict_match(dict_dist_patch_51)
dict_dist_patch_51_sym = create_dict_sym(dict_patch_51)
print(len(dict_dist_patch_51_sym))
print("Abdominal pain" in dict_dist_patch_51_sym.keys())
print("Eye pain" in dict_dist_patch_51_sym.keys())
print("Common cold" in dict_dist_patch_51_sym.keys())

print(list(dict_dist_patch_51_sym.items())[::20])

print(dict_dist_patch_51_sym["Abdominal pain"])
print(dict_dist_patch_51_sym["Eye pain"])

### Closing thoughts:

In any case, as we can see above, the results are not particularly great so there are probably issues with our initial approach to creating or processing the second set of embeddings.

We should proceed with _Embeddings1_ values for now and see if we can pinpoint the issues in the second set of vectors.
