==> ds/notebooks/03_e_ml_clustering.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### NaturDoc - TL BL WT 22-23

# Data clustering:

## HDBSCAN:

## Including Required Code:

### Loading Embeddings Data:


symptoms_embeddings = pd.read_csv("../data/embeddings/word_embeddings_dataframe.csv")

dict_symptom = symptoms_embeddings["Symptom"].to_dict()

#### Transforming:

test_list_1 = symptoms_embeddings.loc[0, "Embedding1"].replace("\n", "").replace("[", "").replace("]", "").split(" ")
test_list_2 = symptoms_embeddings.loc[0, "Embedding2"].replace("\n", "").replace("[", "").replace("]", "").split(" ")

Removing all empty strings:

test_list_1 = [x for x in test_list_1 if x]
test_list_2 = [x for x in test_list_2 if x]
test_list_1[:5]

### Creating useable dataframes:

#### Embedding 1 column:

First, transform content of rows from strings to lists:

symptoms_embeddings.loc[:, "Embedding1"].str.replace("[", "", regex=True).replace("]", "", regex=True).str.split(" ")[1][:10]

def listify_df_values(df_series: pd.Series):
    df_series = df_series.str.replace("\n", "", regex=True)
    df_series = df_series.str.replace("[", "", regex=True).replace("]", "", regex=True)
    df_series = df_series.str.split(" ")
    return df_series

embeddings1_series = listify_df_values(symptoms_embeddings.loc[:, "Embedding1"])

embeddings1_series = embeddings1_series.apply(lambda row: [val for val in row if val])

https://stackoverflow.com/questions/67442107/pandas-expand-explode-dataframe-horizontally

embeddings1_df = pd.DataFrame(embeddings1_series)
embeddings1_df.head(5)

embeddings1_df = pd.concat(
    [embeddings1_df[c].apply(pd.Series).add_prefix(c + "_") for c in embeddings1_df], axis=1
)

embeddings1_df.head(5)

Wrong data type:

embeddings1_df = embeddings1_df.apply(pd.to_numeric, errors='coerce')

### Code to Generate Distance Matrix:

import math
from scipy.spatial import distance_matrix

def generate_distance_matrix(df : pd.DataFrame,
                distance_metric : str = "euclidean") -> pd.DataFrame: # 2.5k x 2.5k
    if distance_metric == "manhattan":
        p = 1
    elif distance_metric == "euclidean":
        p = 2
    elif distance_metric == "chebychev":
        p = math.inf
    else:
        p = 2
    dis_matrix = distance_matrix(df.values, df.values, p)
    dis_df = pd.DataFrame(dis_matrix)
    return dis_df

### Code to Generate Dictionaries:

def generate_dict(df_dist: pd.DataFrame,
                threshold: float) -> dict:
    filt = (df_dist[:] > threshold)
    df_filt = df_dist.copy()
    df_filt[filt] = np.nan
    dict_dist = df_filt.to_dict('dict')
    for i, dic in dict_dist.items():
        to_pop = list()
        for key, value in dic.items():
            if np.isnan(value):
                to_pop.append(key)
        for target_key in to_pop:
            dic.pop(target_key)
        dict_dist[i] = dic
    return dict_dist

def generate_dict_match(dict_dist: dict) -> dict:   
    dict_match = dict()

    for key, value in dict_dist.items():
        for sub_key in value.keys():
            if dict_symptom[key] not in dict_match:
                dict_match[dict_symptom[key]] = [dict_symptom[sub_key]]
            else:
                dict_match[dict_symptom[key]] = [*dict_match.get(dict_symptom[key]), dict_symptom[sub_key]]

    return dict_match

activities_symptoms_df = pd.read_csv("../output/activities_symptoms_bool.csv")
activities_symptoms_df.drop(columns="Unnamed: 0", inplace=True)

filt_sym = (activities_symptoms_df["is_symptom"] == 1)
filt_sym_df = activities_symptoms_df[filt_sym]
filt_sym_list = filt_sym_df["symptomName"].values.tolist()

filt_not_act = (activities_symptoms_df["is_activity"] == 0)
filt_not_act_df = activities_symptoms_df[filt_not_act]
filt_not_act_list = filt_not_act_df["symptomName"].values.tolist()

def create_dict_sym(dict_dist):  
    dict_sym = dict()

    for sym, list_sym in dict_dist.items():
        if sym not in filt_sym_list:
            continue
        for sub_sym in list_sym:
            if sub_sym in filt_not_act_list:
                continue
            if sym not in dict_sym:
                dict_sym[sym] = [sub_sym]
            else:
                dict_sym[sym] = [*dict_sym.get(sym), sub_sym]
    
    return dict_sym

### Generating Distance Matrix and Dictionaries:

df_dist_1 = generate_distance_matrix(embeddings1_df)

dict_dist_086 = generate_dict(df_dist_1, 0.86)
dict_086 = generate_dict_match(dict_dist_086)
dict_086_sym = create_dict_sym(dict_086)

## HDBSCAN* Clustering:

This is more for DS research rather than the MVP functionality. The quality of the matching mostly seems to rely on the word embeddings, in any case.

import hdbscan

### Experimenting With Different Parameters and Attributes:

clusterer_distance = hdbscan.HDBSCAN(metric='precomputed')
clusterer_distance.fit(df_dist_1)
print(set(clusterer_distance.labels_))

clusterer_distance = hdbscan.HDBSCAN(min_samples=4, min_cluster_size=5, metric='precomputed')
clusterer_distance.fit(df_dist_1)
print(set(clusterer_distance.labels_))

print(len(clusterer_distance.probabilities_))
print(set(clusterer_distance.probabilities_[::20]))

#### Labels and probabilities:

print(dict_symptom[0])
print(clusterer_distance.labels_[0])
print(clusterer_distance.probabilities_[0])

print(dict_symptom[1])
print(clusterer_distance.labels_[1])
print(clusterer_distance.probabilities_[1])

print(dict_symptom[1594])
print(clusterer_distance.labels_[1594])
print(clusterer_distance.probabilities_[1594])

label_6 = {
    "indices": [],
    "probabilities": []
}
for i, label in enumerate(clusterer_distance.labels_):
    if label == 6:
        label_6["indices"].append(i)
        label_6["probabilities"].append(clusterer_distance.probabilities_[i])

#### Soft labels:

soft_label_6 = list()

for i, index in enumerate(label_6["indices"]):
    if label_6["probabilities"][i] < 1.0:
        soft_label_6.append(index)

len(soft_label_6)

print(dict_symptom[181])
print(clusterer_distance.labels_[181])
print(clusterer_distance.probabilities_[181])

#### Hard labels:

hard_label_6 = list()

for i, index in enumerate(label_6["indices"]):
    if label_6["probabilities"][i] == 1.0:
        hard_label_6.append(index)

len(hard_label_6)

len(label_6["indices"])

Possible to further refine/reduce number of matches by taking the probabilities value into account.

### Our Cluster Function:

The function must therefore run the HDBSCAN clustering algorithm and return a dictionary where the label groups are returned. min_sample and min_cluster_size should be parameters input into the function to quickly receive different results.

If we want, we can include parameter to make a decision whether to use hard/soft clustering filter, and at what threshold. Implementing this would help further refining the matches in case too many labels with too many items included are generated.

Return a dict of matches and also a list of outliers:

def cluster(df_dist: pd.DataFrame, min_sample_n: int = 5, 
                min_cluster_n: int = 5, filter: bool = False,
                min_probability: float = 1.0) -> dict:
    clusterer_distance = hdbscan.HDBSCAN(
        min_samples = min_sample_n, 
        min_cluster_size = min_cluster_n, 
        metric='precomputed')
    clusterer_distance.fit(df_dist)

    # create a dictionary:
    # a key for each label, value are indices

    labels = list(set(clusterer_distance.labels_))

    label_dict = dict()

    for label in labels:
        label_dict[label] = {}

    for i, assigned_label in enumerate(clusterer_distance.labels_):
        label_dict[assigned_label][i] = clusterer_distance.probabilities_[i]
    
    # create a dictionary:
    # keys are indices from 0 to end, values are all indices from 
    # the same label

    match_dict = dict()
    outliers = list()

    for i in range(len(clusterer_distance.labels_)):
        for label, label_group in label_dict.items():
            if label == -1 and i in label_group:
                outliers.append(i)
                continue
            if i in label_group:
                match_dict[i] = label_group

    return match_dict, outliers, labels

cluster_dict, outliers, labels = cluster(df_dist_1)

labels

Outliers:

print(len(outliers))
outliers[:5]

A dictionary containing similar values:

print(str(cluster_dict)[:200] + "...")

#### Testing out different values:

We have 422 symptoms in our symptom list.

len(filt_sym_list)

cluster_dict, outliers, labels = cluster(df_dist_1, min_sample_n = 2)

labels

print(len(outliers))
outliers[:5]

print(len(cluster_dict[2]))
print(str(cluster_dict)[:200] + "...")

cluster_dict, outliers, labels = cluster(df_dist_1, min_sample_n = 1)
len(labels)

cluster_dict, outliers, labels = cluster(df_dist_1, min_sample_n = 3,
            min_cluster_n = 4)
len(labels)

This is the lowest we can go for min sample and min cluster:

cluster_dict, outliers, labels = cluster(df_dist_1, min_sample_n = 1,
            min_cluster_n = 2)
len(labels)

len(outliers)

print(str(cluster_dict)[:200] + "...")

cluster_sym_dict = generate_dict_match(cluster_dict)

print(str(cluster_sym_dict)[:200] + "...")

final = create_dict_sym(cluster_sym_dict)

print(len(final))
print(str(final)[:200] + "...")

#### Comparison With Dictionary Generated With Distance Matrix and Distance Threshold:

The dictionaries generated with HDBSCAN* are much more restrictive:

print(len(dict_086_sym))
list(dict_086_sym.items())[:5]

print("Abdominal pain" in final.keys())
print(final["Eye pain"])
print("Common cold" in final.keys())

print(dict_086_sym["Abdominal pain"])
print(dict_086_sym["Eye pain"])
print(dict_086_sym["Common cold"])

The results are not as useful, so we will explore more options provided by the HDBSCAN* library in hopes of creating more appropriate clusters.

### Cluster Function with Leaf Clustering:
From the HDBSCAN* documentation:

"_HDBSCAN supports an extra parameter_ cluster_selection_method _to determine how it selects flat clusters from the cluster tree hierarchy. The default method is 'eom' for Excess of Mass, the algorithm described in How HDBSCAN Works. This is not always the most desireable approach to cluster selection. If you are more interested in having small homogeneous clusters then you may find Excess of Mass has a tendency to pick one or two large clusters and then a number of small extra clusters. In this situation you may be tempted to recluster just the data in the single large cluster. Instead, a better option is to select 'leaf' as a cluster selection method. This will select leaf nodes from the tree, producing many small homogeneous clusters. Note that you can still get variable density clusters via this method, and it is also still possible to get large clusters, but there will be a tendency to produce a more fine grained clustering than Excess of Mass can provide._"

def leaf_cluster(df_dist: pd.DataFrame, min_sample_n: int = 5, 
                min_cluster_n: int = 5, filter: bool = False,
                min_probability: float = 1.0) -> dict:
    clusterer_distance = hdbscan.HDBSCAN(
        min_samples = min_sample_n, 
        min_cluster_size = min_cluster_n, 
        cluster_selection_method = 'leaf',
        metric='precomputed')
    clusterer_distance.fit(df_dist)

    # create a dictionary:
    # a key for each label, value are indices

    labels = list(set(clusterer_distance.labels_))

    label_dict = dict()

    for label in labels:
        label_dict[label] = {}

    for i, assigned_label in enumerate(clusterer_distance.labels_):
        label_dict[assigned_label][i] = clusterer_distance.probabilities_[i]
    
    # create a dictionary:
    # keys are indices from 0 to end, values are all indices from 
    # the same label

    match_dict = dict()
    outliers = list()

    for i in range(len(clusterer_distance.labels_)):
        for label, label_group in label_dict.items():
            if label == -1 and i in label_group:
                outliers.append(i)
                continue
            if i in label_group:
                match_dict[i] = label_group

    return match_dict, outliers, labels

As a reminder, previous results:

* 559 labels
* 835 outliers
* 255 symptoms matched to activities in the final dictionary
* "Abdominal pain", "Common cold" not included amongst others

cluster_dict, outliers, labels = leaf_cluster(df_dist_1, min_sample_n = 1,
            min_cluster_n = 2)
print(len(labels), "labels")
print(len(outliers), "outliers")
cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict), "symptoms matched to activities in the final dictionary ")
list(final_cluster_dict.items())[:5]

print("Abdominal pain" in final_cluster_dict.keys())
print(final_cluster_dict["Eye pain"])
print("Common cold" in final_cluster_dict.keys())

Taking a closer look at the dictionary containing all label groups:

cluster_sym_dict = generate_dict_match(cluster_dict)
list(cluster_sym_dict.items())[:5]

The results for leaf clustering are overall quite similar: while a few more labels were created, and less data points labelled as outliers, the dictionary matching only to symptoms actually ended up containing a few less entries. "Common cold" and "Abdominal pain" are still not included.

### Cluster Function with Soft Clustering:

It is important to note that, if we wish to use the soft clustering we should use the prediction_data=True option for HDBSCAN. This will ensure we generate the extra data required that will allow soft clustering to work. This __cannot__ generate prediction data for non-vectorspace inputs - access to the source data rather than mere distances is required!

Accordingly, we should use the original embeddings dataframe and likewise remove the metric = 'precomputed' parameter:

    clusterer_distance = hdbscan.HDBSCAN(
        min_samples = 1, 
        min_cluster_size = 2, 
        prediction_data= True)
    clusterer_distance.fit(df_dist_1)

There is no -1 outliers label, so one less than .labels_ returns

#### Soft Clustering Function:

For each list in soft clusters: 

* label is argmax 

* if the value at argmax index position is above threshold => add to label group

* below threshold: add to outlier dictionary, but label as key

def soft_cluster(df_vect: pd.DataFrame, min_sample_n: int = 5, 
                min_cluster_n: int = 5, threshold: float = 0.2025,
                filter: bool = False,
                min_probability: float = 1.0) -> dict:

    threshold = threshold / 100

    clusterer_distance = hdbscan.HDBSCAN(
        min_samples = min_sample_n, 
        min_cluster_size = min_cluster_n, 
        prediction_data= True)

    clusterer_distance.fit(df_vect)

    soft_clusters = hdbscan.all_points_membership_vectors(clusterer_distance)

    # create a dictionary:
    # a key for each label, value are indices

    labels = range(len(soft_clusters[0]))

    label_dict = dict()
    label_dict[-1] = {}

    for label in labels:
        label_dict[label] = {}

    for i, membership_vectors in enumerate(soft_clusters):
        assigned_label = np.argmax(membership_vectors)
        membership_vector = membership_vectors[assigned_label]
        if membership_vector >= threshold:
            label_dict[assigned_label][i] = membership_vector
        else:
            label_dict[-1][i] = {assigned_label: membership_vector}
    
    # create a dictionary:
    # keys are indices from 0 to end, values are all indices from 
    # the same label

    match_dict = dict()
    outliers = list()

    for i in range(len(soft_clusters)):
        for label, label_group in label_dict.items():
            if label == -1 and i in label_group:
                outliers.append(label_group)
                continue
            if i in label_group:
                match_dict[i] = label_group

    return match_dict, outliers, labels

##### For verification purposes, we are using the basic threshold and trying with distance matrix to check results:

cluster_dict, outliers, labels = soft_cluster(df_dist_1, min_sample_n = 1,
            min_cluster_n = 2)
print(len(labels))
print(len(outliers))

print(str(outliers)[:200] + "...")

cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict))
list(final_cluster_dict.items())[:5]

print("Abdominal pain" in final_cluster_dict.keys())
print("Eye pain" in final_cluster_dict.keys())
print("Common cold" in final_cluster_dict.keys())

final_cluster_dict["Eye pain"]

Results seem somewhat promising (reduction in outliers), but the matches seem a little scrambled. However, this test run used a distance matrix as input, which should have caused issues. 

##### Using Embeddings Dataframe for Soft Clustering:

##### Default Threshold: 

For now, using the default threshold again (which is a little lower than the max membership value discovered amongst previous outliers):

cluster_dict, outliers, labels = soft_cluster(embeddings1_df, min_sample_n = 1,
            min_cluster_n = 2)
print(len(labels), "labels")
print(len(outliers), "outliers")

print(str(outliers)[:200] + "...")

cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict), "symptoms matched in the symptom-activity dictionary")
list(final_cluster_dict.items())[:5]

print("Abdominal pain" in final_cluster_dict.keys())
print("Eye pain" in final_cluster_dict.keys())
print("Common cold" in final_cluster_dict.keys())

##### Threshold 0.202:

cluster_dict, outliers, labels = soft_cluster(embeddings1_df, min_sample_n = 1,
            min_cluster_n = 2, threshold = 0.202)
print(len(labels), "labels")
print(len(outliers), "outliers")
cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict), "symptoms matched in the symptom-activity dictionary")
print("Abdominal pain" in final_cluster_dict.keys())
print("Eye pain" in final_cluster_dict.keys())
print("Common cold" in final_cluster_dict.keys())

list(final_cluster_dict.items())[:5]

##### Threshold 0.1:

cluster_dict, outliers, labels = soft_cluster(embeddings1_df, min_sample_n = 1,
            min_cluster_n = 2, threshold = 0.1)
print(len(labels), "labels")
print(len(outliers), "outliers")
cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict), "symptoms matched in the symptom-activity dictionary")
print("Abdominal pain" in final_cluster_dict.keys())
print("Eye pain" in final_cluster_dict.keys())
print("Common cold" in final_cluster_dict.keys())

list(final_cluster_dict.items())[:5]

final_cluster_dict["Common cold"]

With this threshold, we have an acceptable amount of symptoms matched to activities as recorded in Dr. Duke's database, however they are much too broad and their relation sometimes seems far fetched.

##### Threshold 0.15:

cluster_dict, outliers, labels = soft_cluster(embeddings1_df, min_sample_n = 1,
            min_cluster_n = 2, threshold = 0.15)
print(len(labels), "labels")
print(len(outliers), "outliers")
cluster_sym_dict = generate_dict_match(cluster_dict)
final_cluster_dict = create_dict_sym(cluster_sym_dict)
print(len(final_cluster_dict), "symptoms matched in the symptom-activity dictionary")
print("Abdominal pain" in final_cluster_dict.keys())
print("Eye pain" in final_cluster_dict.keys())
print("Common cold" in final_cluster_dict.keys())

list(final_cluster_dict.items())[:5]

final_cluster_dict["Common cold"]

The soft clustering results continue to seem disappointing. While the length of the final dictionary is alright, and certain matches seem promising, aberrant matches such as "Abdominal pain" matching to "Ankle" continue to be present and cast a shadow of doubt over the results in total.

## Looking Ahead:

The final notebook will deal with visualising the data we have and hopefully yield some more insights into why the clusters sometimes seem so mismatched.
